% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorTypeAggregatedList}
\alias{AcceleratorTypeAggregatedList}
\title{AcceleratorTypeAggregatedList Object}
\usage{
AcceleratorTypeAggregatedList(AcceleratorTypeAggregatedList.items = NULL,
  id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{AcceleratorTypeAggregatedList.items}{The \link{AcceleratorTypeAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{[Output Only] A map of scoped accelerator type lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
AcceleratorTypeAggregatedList object
}
\description{
AcceleratorTypeAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AcceleratorTypeAggregatedList functions: \code{\link{AcceleratorTypeAggregatedList.items}}
}
