% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorTypeList}
\alias{AcceleratorTypeList}
\title{AcceleratorTypeList Object}
\usage{
AcceleratorTypeList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{A list of AcceleratorType resources}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
AcceleratorTypeList object
}
\description{
AcceleratorTypeList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of accelerator types.
}
