% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorTypesScopedList}
\alias{AcceleratorTypesScopedList}
\title{AcceleratorTypesScopedList Object}
\usage{
AcceleratorTypesScopedList(AcceleratorTypesScopedList.warning = NULL,
  AcceleratorTypesScopedList.warning.data = NULL, acceleratorTypes = NULL,
  warning = NULL)
}
\arguments{
\item{AcceleratorTypesScopedList.warning}{The \link{AcceleratorTypesScopedList.warning} object or list of objects}

\item{AcceleratorTypesScopedList.warning.data}{The \link{AcceleratorTypesScopedList.warning.data} object or list of objects}

\item{acceleratorTypes}{[Output Only] List of accelerator types contained in this scope}

\item{warning}{[Output Only] An informational warning that appears when the accelerator types list is empty}
}
\value{
AcceleratorTypesScopedList object
}
\description{
AcceleratorTypesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AcceleratorTypesScopedList functions: \code{\link{AcceleratorTypesScopedList.warning.data}},
  \code{\link{AcceleratorTypesScopedList.warning}}
}
