% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AcceleratorTypesScopedList.warning}
\alias{AcceleratorTypesScopedList.warning}
\title{AcceleratorTypesScopedList.warning Object}
\usage{

  AcceleratorTypesScopedList.warning(AcceleratorTypesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{AcceleratorTypesScopedList.warning.data}{The \link{AcceleratorTypesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
AcceleratorTypesScopedList.warning object
}
\description{
AcceleratorTypesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] An informational warning that appears when the accelerator types list is empty.
}
\seealso{
Other AcceleratorTypesScopedList functions: \code{\link{AcceleratorTypesScopedList.warning.data}},
  \code{\link{AcceleratorTypesScopedList}}
}
