% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AccessConfig}
\alias{AccessConfig}
\title{AccessConfig Object}
\usage{
AccessConfig(name = NULL, natIP = NULL, networkTier = NULL,
  publicDnsName = NULL, publicPtrDomainName = NULL, setPublicDns = NULL,
  setPublicPtr = NULL, type = NULL)
}
\arguments{
\item{name}{Name of this access configuration}

\item{natIP}{An external IP address associated with this instance}

\item{networkTier}{This signifies the networking tier used for configuring this access configuration and can only take the following values: PREMIUM , SELECT}

\item{publicDnsName}{[Output Only] The public DNS domain name for the instance}

\item{publicPtrDomainName}{The DNS domain name for the public PTR record}

\item{setPublicDns}{Specifies whether a public DNS ?A? record should be created for the external IP address of this access configuration}

\item{setPublicPtr}{Specifies whether a public DNS ?PTR? record should be created to map the external IP address of the instance to a DNS domain name}

\item{type}{The type of configuration}
}
\value{
AccessConfig object
}
\description{
AccessConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access configuration attached to an instance's network interface. Only one access config per instance is supported.
}
\seealso{
Other AccessConfig functions: \code{\link{instances.addAccessConfig}},
  \code{\link{instances.updateAccessConfig}}
}
