% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Address}
\alias{Address}
\title{Address Object}
\usage{
Address(Address.labels = NULL, address = NULL, creationTimestamp = NULL,
  description = NULL, id = NULL, ipVersion = NULL,
  labelFingerprint = NULL, labels = NULL, name = NULL,
  networkTier = NULL, region = NULL, selfLink = NULL, status = NULL,
  users = NULL)
}
\arguments{
\item{Address.labels}{The \link{Address.labels} object or list of objects}

\item{address}{The static external IP address represented by this resource}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{ipVersion}{The IP Version that will be used by this address}

\item{labelFingerprint}{A fingerprint for the labels being applied to this Address, which is essentially a hash of the labels set used for optimistic locking}

\item{labels}{Labels to apply to this Address resource}

\item{name}{Name of the resource}

\item{networkTier}{This signifies the networking tier used for configuring this Address and can only take the following values: PREMIUM , SELECT}

\item{region}{[Output Only] URL of the region where the regional address resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] The status of the address, which can be either IN_USE or RESERVED}

\item{users}{[Output Only] The URLs of the resources that are using this address}
}
\value{
Address object
}
\description{
Address Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A reserved address resource.
}
\seealso{
Other Address functions: \code{\link{Address.labels}},
  \code{\link{addresses.insert}},
  \code{\link{globalAddresses.insert}}
}
