% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AddressAggregatedList.items}
\alias{AddressAggregatedList.items}
\title{AddressAggregatedList.items Object}
\usage{
AddressAggregatedList.items()
}
\value{
AddressAggregatedList.items object
}
\description{
AddressAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped address lists.
}
\seealso{
Other AddressAggregatedList functions: \code{\link{AddressAggregatedList}}
}
