% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AddressAggregatedList}
\alias{AddressAggregatedList}
\title{AddressAggregatedList Object}
\usage{
AddressAggregatedList(AddressAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{AddressAggregatedList.items}{The \link{AddressAggregatedList.items} object or list of objects}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{[Output Only] A map of scoped address lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
AddressAggregatedList object
}
\description{
AddressAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AddressAggregatedList functions: \code{\link{AddressAggregatedList.items}}
}

