% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AliasIpRange}
\alias{AliasIpRange}
\title{AliasIpRange Object}
\usage{
AliasIpRange(ipCidrRange = NULL, subnetworkRangeName = NULL)
}
\arguments{
\item{ipCidrRange}{The IP CIDR range represented by this alias IP range}

\item{subnetworkRangeName}{Optional subnetwork secondary range name specifying the secondary range from which to allocate the IP CIDR range for this alias IP range}
}
\value{
AliasIpRange object
}
\description{
AliasIpRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An alias IP range attached to an instance's network interface.
}
