% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AuditConfig}
\alias{AuditConfig}
\title{AuditConfig Object}
\usage{
AuditConfig(exemptedMembers = NULL, service = NULL)
}
\arguments{
\item{exemptedMembers}{Specifies the identities that are exempted from 'data access' audit logging for the `service` specified above}

\item{service}{Specifies a service that will be enabled for 'data access' audit logging}
}
\value{
AuditConfig object
}
\description{
AuditConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Enables 'data access' audit logging for a service and specifies a list of members that are log-exempted.
}

