% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{autoscalers.insert}
\alias{autoscalers.insert}
\title{Creates an autoscaler in the specified project using the data included in the request.}
\usage{
autoscalers.insert(Autoscaler, project, zone)
}
\arguments{
\item{Autoscaler}{The \link{Autoscaler} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{Name of the zone for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Autoscaler functions: \code{\link{Autoscaler}},
  \code{\link{autoscalers.patch}},
  \code{\link{autoscalers.update}},
  \code{\link{regionAutoscalers.insert}},
  \code{\link{regionAutoscalers.patch}},
  \code{\link{regionAutoscalers.update}}
}

