% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicy}
\alias{AutoscalingPolicy}
\title{AutoscalingPolicy Object}
\usage{
AutoscalingPolicy(coolDownPeriodSec = NULL, cpuUtilization = NULL,
  customMetricUtilizations = NULL, loadBalancingUtilization = NULL,
  maxNumReplicas = NULL, minNumReplicas = NULL, queueBasedScaling = NULL)
}
\arguments{
\item{coolDownPeriodSec}{The number of seconds that the autoscaler should wait before it starts collecting information from a new instance}

\item{cpuUtilization}{Defines the CPU utilization policy that allows the autoscaler to scale based on the average CPU utilization of a managed instance group}

\item{customMetricUtilizations}{Configuration parameters of autoscaling based on a custom metric}

\item{loadBalancingUtilization}{Configuration parameters of autoscaling based on load balancer}

\item{maxNumReplicas}{The maximum number of instances that the autoscaler can scale up to}

\item{minNumReplicas}{The minimum number of replicas that the autoscaler can scale down to}

\item{queueBasedScaling}{Configuration parameters of autoscaling based on queuing system}
}
\value{
AutoscalingPolicy object
}
\description{
AutoscalingPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Autoscaler policy.
}
