% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicyQueueBasedScaling}
\alias{AutoscalingPolicyQueueBasedScaling}
\title{AutoscalingPolicyQueueBasedScaling Object}
\usage{
AutoscalingPolicyQueueBasedScaling(acceptableBacklogPerInstance = NULL,
  cloudPubSub = NULL, singleWorkerThroughputPerSec = NULL)
}
\arguments{
\item{acceptableBacklogPerInstance}{Scaling based on the average number of tasks in the queue per each active instance}

\item{cloudPubSub}{Configuration for Cloud Pub/Sub subscription queue}

\item{singleWorkerThroughputPerSec}{The scaling algorithm will also calculate throughput estimates on its own; if you explicitly provide this value, the autoscaler will take into account your value as well as automatic estimates when deciding how to scale}
}
\value{
AutoscalingPolicyQueueBasedScaling object
}
\description{
AutoscalingPolicyQueueBasedScaling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration parameters of autoscaling based on queuing system.
}

