% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AutoscalingPolicyQueueBasedScalingCloudPubSub}
\alias{AutoscalingPolicyQueueBasedScalingCloudPubSub}
\title{AutoscalingPolicyQueueBasedScalingCloudPubSub Object}
\usage{
AutoscalingPolicyQueueBasedScalingCloudPubSub(subscription = NULL,
  topic = NULL)
}
\arguments{
\item{subscription}{Cloud Pub/Sub subscription used for scaling}

\item{topic}{Cloud Pub/Sub topic used for scaling}
}
\value{
AutoscalingPolicyQueueBasedScalingCloudPubSub object
}
\description{
AutoscalingPolicyQueueBasedScalingCloudPubSub Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration parameters for scaling based on Cloud Pub/Sub subscription queue.
}

