% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendBucket}
\alias{BackendBucket}
\title{BackendBucket Object}
\usage{
BackendBucket(bucketName = NULL, creationTimestamp = NULL,
  description = NULL, enableCdn = NULL, id = NULL, name = NULL,
  selfLink = NULL)
}
\arguments{
\item{bucketName}{Cloud Storage bucket name}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{enableCdn}{If true, enable Cloud CDN for this BackendBucket}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{name}{Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
BackendBucket object
}
\description{
BackendBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A BackendBucket resource. This resource defines a Cloud Storage bucket.
}
\seealso{
Other BackendBucket functions: \code{\link{backendBuckets.insert}},
  \code{\link{backendBuckets.patch}},
  \code{\link{backendBuckets.update}}
}

