% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendService}
\alias{BackendService}
\title{BackendService Object}
\usage{
BackendService(affinityCookieTtlSec = NULL, backendSslPolicy = NULL,
  backends = NULL, connectionDraining = NULL, creationTimestamp = NULL,
  description = NULL, enableCDN = NULL, fingerprint = NULL,
  healthChecks = NULL, iaap = NULL, id = NULL,
  loadBalancingScheme = NULL, name = NULL, port = NULL, portName = NULL,
  protocol = NULL, region = NULL, selfLink = NULL,
  sessionAffinity = NULL, timeoutSec = NULL)
}
\arguments{
\item{affinityCookieTtlSec}{Lifetime of cookies in seconds if session_affinity is GENERATED_COOKIE}

\item{backendSslPolicy}{Backend SSL policies to enforce}

\item{backends}{The list of backends that serve this BackendService}

\item{connectionDraining}{No description}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{enableCDN}{If true, enable Cloud CDN for this BackendService}

\item{fingerprint}{Fingerprint of this resource}

\item{healthChecks}{The list of URLs to the HttpHealthCheck or HttpsHealthCheck resource for health checking this BackendService}

\item{iaap}{No description}

\item{id}{[Output Only] The unique identifier for the resource}

\item{loadBalancingScheme}{No description}

\item{name}{Name of the resource}

\item{port}{Deprecated in favor of portName}

\item{portName}{Name of backend port}

\item{protocol}{The protocol this BackendService uses to communicate with backends}

\item{region}{[Output Only] URL of the region where the regional backend service resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sessionAffinity}{Type of session affinity to use}

\item{timeoutSec}{How many seconds to wait for the backend before considering it a failed request}
}
\value{
BackendService object
}
\description{
BackendService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A BackendService resource. This resource defines a group of backend virtual machines and their serving capacity.
}
\seealso{
Other BackendService functions: \code{\link{backendServices.insert}},
  \code{\link{backendServices.patch}},
  \code{\link{backendServices.update}},
  \code{\link{regionBackendServices.insert}},
  \code{\link{regionBackendServices.patch}},
  \code{\link{regionBackendServices.update}}
}

