% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceAggregatedList.items}
\alias{BackendServiceAggregatedList.items}
\title{BackendServiceAggregatedList.items Object}
\usage{
BackendServiceAggregatedList.items()
}
\value{
BackendServiceAggregatedList.items object
}
\description{
BackendServiceAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped BackendService lists.
}
\seealso{
Other BackendServiceAggregatedList functions: \code{\link{BackendServiceAggregatedList}}
}

