% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceCdnPolicy}
\alias{BackendServiceCdnPolicy}
\title{BackendServiceCdnPolicy Object}
\usage{
BackendServiceCdnPolicy(cacheKeyPolicy = NULL, signedUrlKeyNames = NULL,
  signedUrlTtlSec = NULL)
}
\arguments{
\item{cacheKeyPolicy}{The CacheKeyPolicy for this CdnPolicy}

\item{signedUrlKeyNames}{[Output Only] Names of the keys currently configured for Cloud CDN Signed URL on this backend service}

\item{signedUrlTtlSec}{Number of seconds up to which the response to a signed URL request will be cached in the CDN}
}
\value{
BackendServiceCdnPolicy object
}
\description{
BackendServiceCdnPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing Cloud CDN configuration for a backend service.
}
