% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceGroupHealth}
\alias{BackendServiceGroupHealth}
\title{BackendServiceGroupHealth Object}
\usage{
BackendServiceGroupHealth(healthStatus = NULL)
}
\arguments{
\item{healthStatus}{No description}
}
\value{
BackendServiceGroupHealth object
}
\description{
BackendServiceGroupHealth Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
