% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServiceIAAP}
\alias{BackendServiceIAAP}
\title{BackendServiceIAAP Object}
\usage{
BackendServiceIAAP(enabled = NULL, oauth2ClientId = NULL,
  oauth2ClientSecret = NULL, oauth2ClientSecretSha256 = NULL)
}
\arguments{
\item{enabled}{No description}

\item{oauth2ClientId}{No description}

\item{oauth2ClientSecret}{No description}

\item{oauth2ClientSecretSha256}{No description}
}
\value{
BackendServiceIAAP object
}
\description{
BackendServiceIAAP Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identity Aware Application Proxy (Cloud Gatekeeper)
}

