% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendServicesScopedList.warning}
\alias{BackendServicesScopedList.warning}
\title{BackendServicesScopedList.warning Object}
\usage{

  BackendServicesScopedList.warning(BackendServicesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{BackendServicesScopedList.warning.data}{The \link{BackendServicesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
BackendServicesScopedList.warning object
}
\description{
BackendServicesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of backend services when the list is empty.
}
\seealso{
Other BackendServicesScopedList functions: \code{\link{BackendServicesScopedList.warning.data}},
  \code{\link{BackendServicesScopedList}}
}

