% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{BackendSSLPolicy}
\alias{BackendSSLPolicy}
\title{BackendSSLPolicy Object}
\usage{
BackendSSLPolicy(pinnedPeerCertificates = NULL)
}
\arguments{
\item{pinnedPeerCertificates}{List of PEM-encoded peer certificates, from which the public keys are extracted for authenticating the backend service}
}
\value{
BackendSSLPolicy object
}
\description{
BackendSSLPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing backend SSL policies.
}

