% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CacheKeyPolicy}
\alias{CacheKeyPolicy}
\title{CacheKeyPolicy Object}
\usage{
CacheKeyPolicy(includeHost = NULL, includeProtocol = NULL,
  includeQueryString = NULL, queryStringBlacklist = NULL,
  queryStringWhitelist = NULL)
}
\arguments{
\item{includeHost}{If true, requests to different hosts will be cached separately}

\item{includeProtocol}{If true, http and https requests will be cached separately}

\item{includeQueryString}{If true, include query string parameters in the cache key according to query_string_whitelist and query_string_blacklist}

\item{queryStringBlacklist}{Names of query string parameters to exclude in cache keys}

\item{queryStringWhitelist}{Names of query string parameters to include in cache keys}
}
\value{
CacheKeyPolicy object
}
\description{
CacheKeyPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing what to include in the cache key for a request for Cloud CDN.
}
