% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Commitment}
\alias{Commitment}
\title{Commitment Object}
\usage{
Commitment(creationTimestamp = NULL, description = NULL,
  endTimestamp = NULL, id = NULL, name = NULL, plan = NULL,
  region = NULL, resources = NULL, selfLink = NULL,
  startTimestamp = NULL, status = NULL, statusMessage = NULL,
  zone = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{endTimestamp}{[Output Only] Commitment end time in RFC3339 text format}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{plan}{The plan for this commitment, which determines duration and discount rate}

\item{region}{[Output Only] URL of the region where this commitment may be used}

\item{resources}{List of commitment amounts for particular resources}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{startTimestamp}{[Output Only] Commitment start time in RFC3339 text format}

\item{status}{[Output Only] Status of the commitment with regards to eventual expiration (each commitment has an end-date defined)}

\item{statusMessage}{[Output Only] An optional, human-readable explanation of the status}

\item{zone}{[Output Only] URL of the zone where this commitment may be used}
}
\value{
Commitment object
}
\description{
Commitment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A usage-commitment with a start / end time. Users create commitments for particular resources (e.g. memory). Actual usage is first deducted from available commitments made prior, perhaps at a reduced price (as laid out in the commitment).
}
\seealso{
Other Commitment functions: \code{\link{commitments.insert}}
}
