% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CommitmentAggregatedList.items}
\alias{CommitmentAggregatedList.items}
\title{CommitmentAggregatedList.items Object}
\usage{
CommitmentAggregatedList.items()
}
\value{
CommitmentAggregatedList.items object
}
\description{
CommitmentAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Commitments by scope.
}
\seealso{
Other CommitmentAggregatedList functions: \code{\link{CommitmentAggregatedList}}
}
