% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CommitmentAggregatedList}
\alias{CommitmentAggregatedList}
\title{CommitmentAggregatedList Object}
\usage{
CommitmentAggregatedList(CommitmentAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{CommitmentAggregatedList.items}{The \link{CommitmentAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{Commitments by scope}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
CommitmentAggregatedList object
}
\description{
CommitmentAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CommitmentAggregatedList functions: \code{\link{CommitmentAggregatedList.items}}
}
