% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CommitmentsScopedList}
\alias{CommitmentsScopedList}
\title{CommitmentsScopedList Object}
\usage{
CommitmentsScopedList(CommitmentsScopedList.warning = NULL,
  CommitmentsScopedList.warning.data = NULL, commitments = NULL,
  warning = NULL)
}
\arguments{
\item{CommitmentsScopedList.warning}{The \link{CommitmentsScopedList.warning} object or list of objects}

\item{CommitmentsScopedList.warning.data}{The \link{CommitmentsScopedList.warning.data} object or list of objects}

\item{commitments}{[Output Only] List of commitments contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of commitments when the list is empty}
}
\value{
CommitmentsScopedList object
}
\description{
CommitmentsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CommitmentsScopedList functions: \code{\link{CommitmentsScopedList.warning.data}},
  \code{\link{CommitmentsScopedList.warning}}
}
