% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CommitmentsScopedList.warning}
\alias{CommitmentsScopedList.warning}
\title{CommitmentsScopedList.warning Object}
\usage{
CommitmentsScopedList.warning(CommitmentsScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{CommitmentsScopedList.warning.data}{The \link{CommitmentsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
CommitmentsScopedList.warning object
}
\description{
CommitmentsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of commitments when the list is empty.
}
\seealso{
Other CommitmentsScopedList functions: \code{\link{CommitmentsScopedList.warning.data}},
  \code{\link{CommitmentsScopedList}}
}
