% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ConnectionDraining}
\alias{ConnectionDraining}
\title{ConnectionDraining Object}
\usage{
ConnectionDraining(drainingTimeoutSec = NULL)
}
\arguments{
\item{drainingTimeoutSec}{Time for which instance will be drained (not accept new connections, but still work to finish started)}
}
\value{
ConnectionDraining object
}
\description{
ConnectionDraining Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message containing connection draining configuration.
}

