% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{CustomerEncryptionKey}
\alias{CustomerEncryptionKey}
\title{CustomerEncryptionKey Object}
\usage{
CustomerEncryptionKey(rawKey = NULL, rsaEncryptedKey = NULL,
  sha256 = NULL)
}
\arguments{
\item{rawKey}{Specifies a 256-bit customer-supplied encryption key, encoded in RFC 4648 base64 to either encrypt or decrypt this resource}

\item{rsaEncryptedKey}{Specifies an RFC 4648 base64 encoded, RSA-wrapped 2048-bit customer-supplied encryption key to either encrypt or decrypt this resource}

\item{sha256}{[Output only] The RFC 4648 base64 encoded SHA-256 hash of the customer-supplied encryption key that protects this resource}
}
\value{
CustomerEncryptionKey object
}
\description{
CustomerEncryptionKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a customer-supplied encryption key
}
