% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskAggregatedList.items}
\alias{DiskAggregatedList.items}
\title{DiskAggregatedList.items Object}
\usage{
DiskAggregatedList.items()
}
\value{
DiskAggregatedList.items object
}
\description{
DiskAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped disk lists.
}
\seealso{
Other DiskAggregatedList functions: \code{\link{DiskAggregatedList}}
}

