% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskList}
\alias{DiskList}
\title{DiskList Object}
\usage{
DiskList(id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{[Output Only] A list of persistent disks}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
DiskList object
}
\description{
DiskList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of Disk resources.
}

