% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{disks.createSnapshot}
\alias{disks.createSnapshot}
\title{Creates a snapshot of a specified persistent disk.}
\usage{
disks.createSnapshot(Snapshot, project, zone, disk, guestFlush = NULL,
  requestId = NULL)
}
\arguments{
\item{Snapshot}{The \link{Snapshot} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{disk}{Name of the persistent disk to snapshot}

\item{guestFlush}{}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other Snapshot functions: \code{\link{Snapshot.labels}},
  \code{\link{Snapshot}},
  \code{\link{regionDisks.createSnapshot}}
}
