% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DisksResizeRequest}
\alias{DisksResizeRequest}
\title{DisksResizeRequest Object}
\usage{
DisksResizeRequest(sizeGb = NULL)
}
\arguments{
\item{sizeGb}{The new size of the persistent disk, which is specified in GB}
}
\value{
DisksResizeRequest object
}
\description{
DisksResizeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DisksResizeRequest functions: \code{\link{disks.resize}}
}

