% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DisksScopedList}
\alias{DisksScopedList}
\title{DisksScopedList Object}
\usage{
DisksScopedList(DisksScopedList.warning = NULL,
  DisksScopedList.warning.data = NULL, disks = NULL, warning = NULL)
}
\arguments{
\item{DisksScopedList.warning}{The \link{DisksScopedList.warning} object or list of objects}

\item{DisksScopedList.warning.data}{The \link{DisksScopedList.warning.data} object or list of objects}

\item{disks}{[Output Only] List of disks contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of disks when the list is empty}
}
\value{
DisksScopedList object
}
\description{
DisksScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DisksScopedList functions: \code{\link{DisksScopedList.warning.data}},
  \code{\link{DisksScopedList.warning}}
}

