% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DisksScopedList.warning}
\alias{DisksScopedList.warning}
\title{DisksScopedList.warning Object}
\usage{
DisksScopedList.warning(DisksScopedList.warning.data = NULL, code = NULL,
  data = NULL, message = NULL)
}
\arguments{
\item{DisksScopedList.warning.data}{The \link{DisksScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
DisksScopedList.warning object
}
\description{
DisksScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of disks when the list is empty.
}
\seealso{
Other DisksScopedList functions: \code{\link{DisksScopedList.warning.data}},
  \code{\link{DisksScopedList}}
}
