% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskType}
\alias{DiskType}
\title{DiskType Object}
\usage{
DiskType(creationTimestamp = NULL, defaultDiskSizeGb = NULL,
  deprecated = NULL, description = NULL, id = NULL, name = NULL,
  selfLink = NULL, validDiskSize = NULL, zone = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{defaultDiskSizeGb}{[Output Only] Server-defined default disk size in GB}

\item{deprecated}{[Output Only] The deprecation status associated with this disk type}

\item{description}{[Output Only] An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{[Output Only] Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{validDiskSize}{[Output Only] An optional textual description of the valid disk size, such as '10GB-10TB'}

\item{zone}{[Output Only] URL of the zone where the disk type resides}
}
\value{
DiskType object
}
\description{
DiskType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A DiskType resource.
}

