% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DiskTypeAggregatedList.items}
\alias{DiskTypeAggregatedList.items}
\title{DiskTypeAggregatedList.items Object}
\usage{
DiskTypeAggregatedList.items()
}
\value{
DiskTypeAggregatedList.items object
}
\description{
DiskTypeAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped disk type lists.
}
\seealso{
Other DiskTypeAggregatedList functions: \code{\link{DiskTypeAggregatedList}}
}
