% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{DistributionPolicyZoneConfiguration}
\alias{DistributionPolicyZoneConfiguration}
\title{DistributionPolicyZoneConfiguration Object}
\usage{
DistributionPolicyZoneConfiguration(zone = NULL)
}
\arguments{
\item{zone}{URL of the zone where managed instance group is spawning instances (for regional resources)}
}
\value{
DistributionPolicyZoneConfiguration object
}
\description{
DistributionPolicyZoneConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
