% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Firewall.denied}
\alias{Firewall.denied}
\title{Firewall.denied Object}
\usage{
Firewall.denied()
}
\value{
Firewall.denied object
}
\description{
Firewall.denied Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Firewall functions: \code{\link{Firewall.allowed}},
  \code{\link{Firewall}}, \code{\link{firewalls.insert}},
  \code{\link{firewalls.patch}},
  \code{\link{firewalls.update}}
}
