% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{FixedOrPercent}
\alias{FixedOrPercent}
\title{FixedOrPercent Object}
\usage{
FixedOrPercent(calculated = NULL, fixed = NULL, percent = NULL)
}
\arguments{
\item{calculated}{[Output Only] Absolute value calculated based on mode: mode = fixed -> calculated = fixed = percent -> calculated = ceiling(percent/100 * base_value)}

\item{fixed}{fixed must be non-negative}

\item{percent}{percent must belong to [0, 100]}
}
\value{
FixedOrPercent object
}
\description{
FixedOrPercent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Encapsulates numeric value that can be either absolute or relative.
}

