% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRule.labels}
\alias{ForwardingRule.labels}
\title{ForwardingRule.labels Object}
\usage{
ForwardingRule.labels()
}
\value{
ForwardingRule.labels object
}
\description{
ForwardingRule.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this resource. These can be later modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.
}
\seealso{
Other ForwardingRule functions: \code{\link{ForwardingRule}},
  \code{\link{forwardingRules.insert}},
  \code{\link{globalForwardingRules.insert}}
}
