% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ForwardingRule}
\alias{ForwardingRule}
\title{ForwardingRule Object}
\usage{
ForwardingRule(ForwardingRule.labels = NULL, IPAddress = NULL,
  IPProtocol = NULL, backendService = NULL, creationTimestamp = NULL,
  description = NULL, id = NULL, ipVersion = NULL,
  labelFingerprint = NULL, labels = NULL, loadBalancingScheme = NULL,
  name = NULL, network = NULL, networkTier = NULL, portRange = NULL,
  ports = NULL, region = NULL, selfLink = NULL, serviceLabel = NULL,
  serviceName = NULL, subnetwork = NULL, target = NULL)
}
\arguments{
\item{ForwardingRule.labels}{The \link{ForwardingRule.labels} object or list of objects}

\item{IPAddress}{The IP address that this forwarding rule is serving on behalf of}

\item{IPProtocol}{The IP protocol to which this rule applies}

\item{backendService}{This field is not used for external load balancing}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{ipVersion}{The IP Version that will be used by this forwarding rule}

\item{labelFingerprint}{A fingerprint for the labels being applied to this resource, which is essentially a hash of the labels set used for optimistic locking}

\item{labels}{Labels to apply to this resource}

\item{loadBalancingScheme}{This signifies what the ForwardingRule will be used for and can only take the following values: INTERNAL, EXTERNAL The value of INTERNAL means that this will be used for Internal Network Load Balancing (TCP, UDP)}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{network}{This field is not used for external load balancing}

\item{networkTier}{This signifies the networking tier used for configuring this load balancer and can only take the following values: PREMIUM , SELECT}

\item{portRange}{Applicable only when IPProtocol is TCP, UDP, or SCTP, only packets addressed to ports in the specified range will be forwarded to target}

\item{ports}{This field is not used for external load balancing}

\item{region}{[Output Only] URL of the region where the regional forwarding rule resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{serviceLabel}{An optional prefix to the service name for this Forwarding Rule}

\item{serviceName}{[Output Only] The internal fully qualified service name for this Forwarding Rule}

\item{subnetwork}{This field is not used for external load balancing}

\item{target}{The URL of the target resource to receive the matched traffic}
}
\value{
ForwardingRule object
}
\description{
ForwardingRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A ForwardingRule resource. A ForwardingRule resource specifies which pool of target virtual machines to forward a packet to if it matches the given [IPAddress, IPProtocol, portRange] tuple.
}
\seealso{
Other ForwardingRule functions: \code{\link{ForwardingRule.labels}},
  \code{\link{forwardingRules.insert}},
  \code{\link{globalForwardingRules.insert}}
}
