% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{forwardingRules.setTarget}
\alias{forwardingRules.setTarget}
\title{Changes target URL for forwarding rule. The new target should be of the same type as the old target.}
\usage{
forwardingRules.setTarget(TargetReference, project, region, forwardingRule,
  requestId = NULL)
}
\arguments{
\item{TargetReference}{The \link{TargetReference} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{forwardingRule}{Name of the ForwardingRule resource in which target is to be set}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetReference functions: \code{\link{TargetReference}},
  \code{\link{globalForwardingRules.setTarget}},
  \code{\link{targetPools.setBackup}}
}
