% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{GlobalSetLabelsRequest}
\alias{GlobalSetLabelsRequest}
\title{GlobalSetLabelsRequest Object}
\usage{
GlobalSetLabelsRequest(GlobalSetLabelsRequest.labels = NULL,
  labelFingerprint = NULL, labels = NULL)
}
\arguments{
\item{GlobalSetLabelsRequest.labels}{The \link{GlobalSetLabelsRequest.labels} object or list of objects}

\item{labelFingerprint}{The fingerprint of the previous set of labels for this resource, used to detect conflicts}

\item{labels}{A list of labels to apply for this resource}
}
\value{
GlobalSetLabelsRequest object
}
\description{
GlobalSetLabelsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GlobalSetLabelsRequest functions: \code{\link{GlobalSetLabelsRequest.labels}},
  \code{\link{globalAddresses.setLabels}},
  \code{\link{globalForwardingRules.setLabels}},
  \code{\link{images.setLabels}},
  \code{\link{snapshots.setLabels}}
}
