% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{GuestOsFeature}
\alias{GuestOsFeature}
\title{GuestOsFeature Object}
\usage{
GuestOsFeature(type = NULL)
}
\arguments{
\item{type}{The type of supported feature}
}
\value{
GuestOsFeature object
}
\description{
GuestOsFeature Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Features supported by the guest os.
}

