% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Host}
\alias{Host}
\title{Host Object}
\usage{
Host(creationTimestamp = NULL, description = NULL, hostType = NULL,
  id = NULL, instances = NULL, name = NULL, selfLink = NULL,
  status = NULL, statusMessage = NULL, zone = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{[Output Only] An optional textual description of the resource}

\item{hostType}{Full or partial URL of the host type resource to use for this host, in the format: zones/zone/hostTypes/host-type}

\item{id}{[Output Only] The unique identifier for the resource}

\item{instances}{A list of resource URLs to the virtual machine instances in this host}

\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] The status of the host}

\item{statusMessage}{[Output Only] An optional, human-readable explanation of the status}

\item{zone}{[Output Only] The name of the zone where the host resides, such as us-central1-a}
}
\value{
Host object
}
\description{
Host Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Host functions: \code{\link{hosts.insert}}
}
