% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostAggregatedList.items}
\alias{HostAggregatedList.items}
\title{HostAggregatedList.items Object}
\usage{
HostAggregatedList.items()
}
\value{
HostAggregatedList.items object
}
\description{
HostAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped host lists.
}
\seealso{
Other HostAggregatedList functions: \code{\link{HostAggregatedList}}
}
