% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostRule}
\alias{HostRule}
\title{HostRule Object}
\usage{
HostRule(description = NULL, hosts = NULL, pathMatcher = NULL)
}
\arguments{
\item{description}{An optional description of this resource}

\item{hosts}{The list of host patterns to match}

\item{pathMatcher}{The name of the PathMatcher to use to match the path portion of the URL if the hostRule matches the URL's host portion}
}
\value{
HostRule object
}
\description{
HostRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
UrlMaps A host-matching rule for a URL. If matched, will use the named PathMatcher to select the BackendService.
}

