% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostsScopedList}
\alias{HostsScopedList}
\title{HostsScopedList Object}
\usage{
HostsScopedList(HostsScopedList.warning = NULL,
  HostsScopedList.warning.data = NULL, hosts = NULL, warning = NULL)
}
\arguments{
\item{HostsScopedList.warning}{The \link{HostsScopedList.warning} object or list of objects}

\item{HostsScopedList.warning.data}{The \link{HostsScopedList.warning.data} object or list of objects}

\item{hosts}{[Output Only] List of hosts contained in this scope}

\item{warning}{[Output Only] An informational warning that appears when the host list is empty}
}
\value{
HostsScopedList object
}
\description{
HostsScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other HostsScopedList functions: \code{\link{HostsScopedList.warning.data}},
  \code{\link{HostsScopedList.warning}}
}
