% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostsScopedList.warning.data}
\alias{HostsScopedList.warning.data}
\title{HostsScopedList.warning.data Object}
\usage{
HostsScopedList.warning.data()
}
\value{
HostsScopedList.warning.data object
}
\description{
HostsScopedList.warning.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other HostsScopedList functions: \code{\link{HostsScopedList.warning}},
  \code{\link{HostsScopedList}}
}
