% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostsScopedList.warning}
\alias{HostsScopedList.warning}
\title{HostsScopedList.warning Object}
\usage{
HostsScopedList.warning(HostsScopedList.warning.data = NULL, code = NULL,
  data = NULL, message = NULL)
}
\arguments{
\item{HostsScopedList.warning.data}{The \link{HostsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
HostsScopedList.warning object
}
\description{
HostsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] An informational warning that appears when the host list is empty.
}
\seealso{
Other HostsScopedList functions: \code{\link{HostsScopedList.warning.data}},
  \code{\link{HostsScopedList}}
}
