% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostType}
\alias{HostType}
\title{HostType Object}
\usage{
HostType(cpuPlatform = NULL, creationTimestamp = NULL, deprecated = NULL,
  description = NULL, guestCpus = NULL, id = NULL, localSsdGb = NULL,
  memoryMb = NULL, name = NULL, selfLink = NULL, zone = NULL)
}
\arguments{
\item{cpuPlatform}{[Output Only] The CPU platform used by this host type}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{[Output Only] The deprecation status associated with this host type}

\item{description}{[Output Only] An optional textual description of the resource}

\item{guestCpus}{[Output Only] The number of virtual CPUs that are available to the host type}

\item{id}{[Output Only] The unique identifier for the resource}

\item{localSsdGb}{[Output Only] Local SSD available to the host type, defined in GB}

\item{memoryMb}{[Output Only] The amount of physical memory available to the host type, defined in MB}

\item{name}{[Output Only] Name of the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{zone}{[Output Only] The name of the zone where the host type resides, such as us-central1-a}
}
\value{
HostType object
}
\description{
HostType Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Host Type resource.
}
