% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostTypeAggregatedList.items}
\alias{HostTypeAggregatedList.items}
\title{HostTypeAggregatedList.items Object}
\usage{
HostTypeAggregatedList.items()
}
\value{
HostTypeAggregatedList.items object
}
\description{
HostTypeAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] A map of scoped host type lists.
}
\seealso{
Other HostTypeAggregatedList functions: \code{\link{HostTypeAggregatedList}}
}
