% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostTypeAggregatedList}
\alias{HostTypeAggregatedList}
\title{HostTypeAggregatedList Object}
\usage{
HostTypeAggregatedList(HostTypeAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{HostTypeAggregatedList.items}{The \link{HostTypeAggregatedList.items} object or list of objects}

\item{id}{[Output Only] The unique identifier for the resource}

\item{items}{[Output Only] A map of scoped host type lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
HostTypeAggregatedList object
}
\description{
HostTypeAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other HostTypeAggregatedList functions: \code{\link{HostTypeAggregatedList.items}}
}
