% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostTypesScopedList}
\alias{HostTypesScopedList}
\title{HostTypesScopedList Object}
\usage{
HostTypesScopedList(HostTypesScopedList.warning = NULL,
  HostTypesScopedList.warning.data = NULL, hostTypes = NULL,
  warning = NULL)
}
\arguments{
\item{HostTypesScopedList.warning}{The \link{HostTypesScopedList.warning} object or list of objects}

\item{HostTypesScopedList.warning.data}{The \link{HostTypesScopedList.warning.data} object or list of objects}

\item{hostTypes}{[Output Only] List of host types contained in this scope}

\item{warning}{[Output Only] An informational warning that appears when the host types list is empty}
}
\value{
HostTypesScopedList object
}
\description{
HostTypesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other HostTypesScopedList functions: \code{\link{HostTypesScopedList.warning.data}},
  \code{\link{HostTypesScopedList.warning}}
}
