% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostTypesScopedList.warning.data}
\alias{HostTypesScopedList.warning.data}
\title{HostTypesScopedList.warning.data Object}
\usage{
HostTypesScopedList.warning.data()
}
\value{
HostTypesScopedList.warning.data object
}
\description{
HostTypesScopedList.warning.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other HostTypesScopedList functions: \code{\link{HostTypesScopedList.warning}},
  \code{\link{HostTypesScopedList}}
}
