% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HostTypesScopedList.warning}
\alias{HostTypesScopedList.warning}
\title{HostTypesScopedList.warning Object}
\usage{
HostTypesScopedList.warning(HostTypesScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{HostTypesScopedList.warning.data}{The \link{HostTypesScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
HostTypesScopedList.warning object
}
\description{
HostTypesScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] An informational warning that appears when the host types list is empty.
}
\seealso{
Other HostTypesScopedList functions: \code{\link{HostTypesScopedList.warning.data}},
  \code{\link{HostTypesScopedList}}
}
