% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{HTTP2HealthCheck}
\alias{HTTP2HealthCheck}
\title{HTTP2HealthCheck Object}
\usage{
HTTP2HealthCheck(host = NULL, port = NULL, portName = NULL,
  proxyHeader = NULL, requestPath = NULL, response = NULL)
}
\arguments{
\item{host}{The value of the host header in the HTTP/2 health check request}

\item{port}{The TCP port number for the health check request}

\item{portName}{Port name as defined in InstanceGroup#NamedPort#name}

\item{proxyHeader}{Specifies the type of proxy header to append before sending data to the backend, either NONE or PROXY_V1}

\item{requestPath}{The request path of the HTTP/2 health check request}

\item{response}{The string to match anywhere in the first 1024 bytes of the response body}
}
\value{
HTTP2HealthCheck object
}
\description{
HTTP2HealthCheck Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
