% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{httpHealthChecks.testIamPermissions}
\alias{httpHealthChecks.testIamPermissions}
\title{Returns permissions that a caller has on the specified resource.}
\usage{
httpHealthChecks.testIamPermissions(TestPermissionsRequest, project, resource)
}
\arguments{
\item{TestPermissionsRequest}{The \link{TestPermissionsRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{resource}{Name of the resource for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TestPermissionsRequest functions: \code{\link{TestPermissionsRequest}},
  \code{\link{addresses.testIamPermissions}},
  \code{\link{autoscalers.testIamPermissions}},
  \code{\link{backendBuckets.testIamPermissions}},
  \code{\link{backendServices.testIamPermissions}},
  \code{\link{clientSslPolicies.testIamPermissions}},
  \code{\link{commitments.testIamPermissions}},
  \code{\link{disks.testIamPermissions}},
  \code{\link{firewalls.testIamPermissions}},
  \code{\link{forwardingRules.testIamPermissions}},
  \code{\link{globalAddresses.testIamPermissions}},
  \code{\link{globalForwardingRules.testIamPermissions}},
  \code{\link{healthChecks.testIamPermissions}},
  \code{\link{hosts.testIamPermissions}},
  \code{\link{httpsHealthChecks.testIamPermissions}},
  \code{\link{images.testIamPermissions}},
  \code{\link{instanceGroupManagers.testIamPermissions}},
  \code{\link{instanceGroups.testIamPermissions}},
  \code{\link{instanceTemplates.testIamPermissions}},
  \code{\link{instances.testIamPermissions}},
  \code{\link{networks.testIamPermissions}},
  \code{\link{regionAutoscalers.testIamPermissions}},
  \code{\link{regionBackendServices.testIamPermissions}},
  \code{\link{regionDisks.testIamPermissions}},
  \code{\link{regionInstanceGroupManagers.testIamPermissions}},
  \code{\link{regionInstanceGroups.testIamPermissions}},
  \code{\link{routers.testIamPermissions}},
  \code{\link{routes.testIamPermissions}},
  \code{\link{snapshots.testIamPermissions}},
  \code{\link{sslCertificates.testIamPermissions}},
  \code{\link{subnetworks.testIamPermissions}},
  \code{\link{targetHttpProxies.testIamPermissions}},
  \code{\link{targetHttpsProxies.testIamPermissions}},
  \code{\link{targetInstances.testIamPermissions}},
  \code{\link{targetPools.testIamPermissions}},
  \code{\link{targetSslProxies.testIamPermissions}},
  \code{\link{targetTcpProxies.testIamPermissions}},
  \code{\link{targetVpnGateways.testIamPermissions}},
  \code{\link{urlMaps.testIamPermissions}},
  \code{\link{vpnTunnels.testIamPermissions}}
}
