% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{httpsHealthChecks.insert}
\alias{httpsHealthChecks.insert}
\title{Creates a HttpsHealthCheck resource in the specified project using the data included in the request.}
\usage{
httpsHealthChecks.insert(HttpsHealthCheck, project, requestId = NULL)
}
\arguments{
\item{HttpsHealthCheck}{The \link{HttpsHealthCheck} object to pass to this method}

\item{project}{Project ID for this request}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other HttpsHealthCheck functions: \code{\link{HttpsHealthCheck}},
  \code{\link{httpsHealthChecks.patch}},
  \code{\link{httpsHealthChecks.update}}
}
