% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{httpsHealthChecks.patch}
\alias{httpsHealthChecks.patch}
\title{Updates a HttpsHealthCheck resource in the specified project using the data included in the request. This method supports patch semantics.}
\usage{
httpsHealthChecks.patch(HttpsHealthCheck, project, httpsHealthCheck,
  requestId = NULL)
}
\arguments{
\item{HttpsHealthCheck}{The \link{HttpsHealthCheck} object to pass to this method}

\item{project}{Project ID for this request}

\item{httpsHealthCheck}{Name of the HttpsHealthCheck resource to patch}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other HttpsHealthCheck functions: \code{\link{HttpsHealthCheck}},
  \code{\link{httpsHealthChecks.insert}},
  \code{\link{httpsHealthChecks.update}}
}
