% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Image.labels}
\alias{Image.labels}
\title{Image.labels Object}
\usage{
Image.labels()
}
\value{
Image.labels object
}
\description{
Image.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this image. These can be later modified by the setLabels method.
}
\seealso{
Other Image functions: \code{\link{Image.rawDisk}},
  \code{\link{Image}}, \code{\link{images.insert}}
}
