% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(Instance.labels = NULL, canIpForward = NULL, cpuPlatform = NULL,
  creationTimestamp = NULL, description = NULL, disks = NULL,
  guestAccelerators = NULL, host = NULL, id = NULL,
  instanceEncryptionKey = NULL, labelFingerprint = NULL, labels = NULL,
  machineType = NULL, metadata = NULL, minCpuPlatform = NULL,
  name = NULL, networkInterfaces = NULL, scheduling = NULL,
  selfLink = NULL, serviceAccounts = NULL, status = NULL,
  statusMessage = NULL, tags = NULL, zone = NULL)
}
\arguments{
\item{Instance.labels}{The \link{Instance.labels} object or list of objects}

\item{canIpForward}{Allows this instance to send and receive packets with non-matching destination or source IPs}

\item{cpuPlatform}{[Output Only] The CPU platform used by this instance}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{disks}{Array of disks associated with this instance}

\item{guestAccelerators}{No description}

\item{host}{Full or partial URL of the host resource that the instance should be placed on, in the format: zones/zone/hosts/host}

\item{id}{[Output Only] The unique identifier for the resource}

\item{instanceEncryptionKey}{Encrypts or decrypts data for an instance with a customer-supplied encryption key}

\item{labelFingerprint}{A fingerprint for this request, which is essentially a hash of the metadata's contents and used for optimistic locking}

\item{labels}{Labels to apply to this instance}

\item{machineType}{Full or partial URL of the machine type resource to use for this instance, in the format: zones/zone/machineTypes/machine-type}

\item{metadata}{The metadata key/value pairs assigned to this instance}

\item{minCpuPlatform}{Minimum cpu/platform to be used by this instance}

\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{networkInterfaces}{An array of configurations for this interface}

\item{scheduling}{Scheduling options for this instance}

\item{selfLink}{[Output Only] Server-defined URL for this resource}

\item{serviceAccounts}{A list of service accounts, with their specified scopes, authorized for this instance}

\item{status}{[Output Only] The status of the instance}

\item{statusMessage}{[Output Only] An optional, human-readable explanation of the status}

\item{tags}{A list of tags to apply to this instance}

\item{zone}{[Output Only] URL of the zone where the instance resides}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance resource.
}
\seealso{
Other Instance functions: \code{\link{Instance.labels}},
  \code{\link{instances.insert}}
}
