% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupAggregatedList.items}
\alias{InstanceGroupAggregatedList.items}
\title{InstanceGroupAggregatedList.items Object}
\usage{
InstanceGroupAggregatedList.items()
}
\value{
InstanceGroupAggregatedList.items object
}
\description{
InstanceGroupAggregatedList.items Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of scoped instance group lists.
}
\seealso{
Other InstanceGroupAggregatedList functions: \code{\link{InstanceGroupAggregatedList}}
}

