% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManager}
\alias{InstanceGroupManager}
\title{InstanceGroupManager Object}
\usage{
InstanceGroupManager(autoHealingPolicies = NULL, baseInstanceName = NULL,
  creationTimestamp = NULL, currentActions = NULL, description = NULL,
  failoverAction = NULL, fingerprint = NULL, id = NULL,
  instanceGroup = NULL, instanceTemplate = NULL, name = NULL,
  namedPorts = NULL, pendingActions = NULL, region = NULL,
  selfLink = NULL, targetPools = NULL, targetSize = NULL,
  updatePolicy = NULL, versions = NULL, zone = NULL)
}
\arguments{
\item{autoHealingPolicies}{The autohealing policy for this managed instance group}

\item{baseInstanceName}{The base instance name to use for instances in this group}

\item{creationTimestamp}{[Output Only] The creation timestamp for this managed instance group in RFC3339 text format}

\item{currentActions}{[Output Only] The list of instance actions and the number of instances in this managed instance group that are scheduled for each of those actions}

\item{description}{An optional description of this resource}

\item{failoverAction}{The action to perform in case of zone failure (set only for Regional instance group managers)}

\item{fingerprint}{[Output Only] The fingerprint of the resource data}

\item{id}{[Output Only] A unique identifier for this resource type}

\item{instanceGroup}{[Output Only] The URL of the Instance Group resource}

\item{instanceTemplate}{The URL of the instance template that is specified for this managed instance group}

\item{name}{The name of the managed instance group}

\item{namedPorts}{Named ports configured for the Instance Groups complementary to this Instance Group Manager}

\item{pendingActions}{[Output Only] The list of instance actions and the number of instances in this managed instance group that are pending for each of those actions}

\item{region}{[Output Only] The URL of the region where the managed instance group resides (for regional resources)}

\item{selfLink}{[Output Only] The URL for this managed instance group}

\item{targetPools}{The URLs for all TargetPool resources to which instances in the instanceGroup field are added}

\item{targetSize}{The target number of running instances for this managed instance group}

\item{updatePolicy}{The update policy for this managed instance group}

\item{versions}{Versions supported by this IGM}

\item{zone}{[Output Only] The URL of the zone where the managed instance group is located (for zonal resources)}
}
\value{
InstanceGroupManager object
}
\description{
InstanceGroupManager Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance Group Manager resource.
}
\seealso{
Other InstanceGroupManager functions: \code{\link{instanceGroupManagers.insert}},
  \code{\link{instanceGroupManagers.patch}},
  \code{\link{instanceGroupManagers.update}},
  \code{\link{regionInstanceGroupManagers.insert}},
  \code{\link{regionInstanceGroupManagers.patch}},
  \code{\link{regionInstanceGroupManagers.update}}
}

