% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerActionsSummary}
\alias{InstanceGroupManagerActionsSummary}
\title{InstanceGroupManagerActionsSummary Object}
\usage{
InstanceGroupManagerActionsSummary(abandoning = NULL, creating = NULL,
  creatingWithoutRetries = NULL, deleting = NULL, none = NULL,
  recreating = NULL, refreshing = NULL, restarting = NULL,
  verifying = NULL)
}
\arguments{
\item{abandoning}{[Output Only] The total number of instances in the managed instance group that are scheduled to be abandoned}

\item{creating}{[Output Only] The number of instances in the managed instance group that are scheduled to be created or are currently being created}

\item{creatingWithoutRetries}{[Output Only] The number of instances that the managed instance group will attempt to create}

\item{deleting}{[Output Only] The number of instances in the managed instance group that are scheduled to be deleted or are currently being deleted}

\item{none}{[Output Only] The number of instances in the managed instance group that are running and have no scheduled actions}

\item{recreating}{[Output Only] The number of instances in the managed instance group that are scheduled to be recreated or are currently being being recreated}

\item{refreshing}{[Output Only] The number of instances in the managed instance group that are being reconfigured with properties that do not require a restart or a recreate action}

\item{restarting}{[Output Only] The number of instances in the managed instance group that are scheduled to be restarted or are currently being restarted}

\item{verifying}{[Output Only] The number of instances in the managed instance group that are being verified}
}
\value{
InstanceGroupManagerActionsSummary object
}
\description{
InstanceGroupManagerActionsSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
