% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerPendingActionsSummary}
\alias{InstanceGroupManagerPendingActionsSummary}
\title{InstanceGroupManagerPendingActionsSummary Object}
\usage{
InstanceGroupManagerPendingActionsSummary(creating = NULL, deleting = NULL,
  recreating = NULL, restarting = NULL)
}
\arguments{
\item{creating}{[Output Only] The number of instances in the managed instance group that are pending to be created}

\item{deleting}{[Output Only] The number of instances in the managed instance group that are pending to be deleted}

\item{recreating}{[Output Only] The number of instances in the managed instance group that are pending to be recreated}

\item{restarting}{[Output Only] The number of instances in the managed instance group that are pending to be restarted}
}
\value{
InstanceGroupManagerPendingActionsSummary object
}
\description{
InstanceGroupManagerPendingActionsSummary Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
