% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroupManagers.resizeAdvanced}
\alias{instanceGroupManagers.resizeAdvanced}
\title{Resizes the managed instance group with advanced configuration options like disabling creation retries. This is an extended version of the resize method.If you increase the size of the instance group, the group creates new instances using the current instance template. If you decrease the size, the group deletes instances. The resize operation is marked DONE when the resize actions are scheduled even if the group has not yet added or deleted any instances. You must separately verify the status of the creating, creatingWithoutRetries, or deleting actions with the get or listmanagedinstances method.}
\usage{

  instanceGroupManagers.resizeAdvanced(InstanceGroupManagersResizeAdvancedRequest,
  project, zone, instanceGroupManager)
}
\arguments{
\item{InstanceGroupManagersResizeAdvancedRequest}{The \link{InstanceGroupManagersResizeAdvancedRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the managed instance group is located}

\item{instanceGroupManager}{The name of the managed instance group}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupManagersResizeAdvancedRequest functions: \code{\link{InstanceGroupManagersResizeAdvancedRequest}}
}

