% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroupManagers.setTargetPools}
\alias{instanceGroupManagers.setTargetPools}
\title{Modifies the target pools to which all instances in this managed instance group are assigned. The target pools automatically apply to all of the instances in the managed instance group. This operation is marked DONE when you make the request even if the instances have not yet been added to their target pools. The change might take some time to apply to all of the instances in the group depending on the size of the group.}
\usage{

  instanceGroupManagers.setTargetPools(InstanceGroupManagersSetTargetPoolsRequest,
  project, zone, instanceGroupManager, requestId = NULL)
}
\arguments{
\item{InstanceGroupManagersSetTargetPoolsRequest}{The \link{InstanceGroupManagersSetTargetPoolsRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the managed instance group is located}

\item{instanceGroupManager}{The name of the managed instance group}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupManagersSetTargetPoolsRequest functions: \code{\link{InstanceGroupManagersSetTargetPoolsRequest}}
}
