% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersListManagedInstancesResponse}
\alias{InstanceGroupManagersListManagedInstancesResponse}
\title{InstanceGroupManagersListManagedInstancesResponse Object}
\usage{
InstanceGroupManagersListManagedInstancesResponse(managedInstances = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{managedInstances}{[Output Only] The list of instances in the managed instance group}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}
}
\value{
InstanceGroupManagersListManagedInstancesResponse object
}
\description{
InstanceGroupManagersListManagedInstancesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
