% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersResizeAdvancedRequest}
\alias{InstanceGroupManagersResizeAdvancedRequest}
\title{InstanceGroupManagersResizeAdvancedRequest Object}
\usage{
InstanceGroupManagersResizeAdvancedRequest(noCreationRetries = NULL,
  targetSize = NULL)
}
\arguments{
\item{noCreationRetries}{If this flag is true, the managed instance group attempts to create all instances initiated by this resize request only once}

\item{targetSize}{The number of running instances that the managed instance group should maintain at any given time}
}
\value{
InstanceGroupManagersResizeAdvancedRequest object
}
\description{
InstanceGroupManagersResizeAdvancedRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersResizeAdvancedRequest functions: \code{\link{instanceGroupManagers.resizeAdvanced}}
}

