% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersScopedList}
\alias{InstanceGroupManagersScopedList}
\title{InstanceGroupManagersScopedList Object}
\usage{

  InstanceGroupManagersScopedList(InstanceGroupManagersScopedList.warning = NULL,
  InstanceGroupManagersScopedList.warning.data = NULL,
  instanceGroupManagers = NULL, warning = NULL)
}
\arguments{
\item{InstanceGroupManagersScopedList.warning}{The \link{InstanceGroupManagersScopedList.warning} object or list of objects}

\item{InstanceGroupManagersScopedList.warning.data}{The \link{InstanceGroupManagersScopedList.warning.data} object or list of objects}

\item{instanceGroupManagers}{[Output Only] The list of managed instance groups that are contained in the specified project and zone}

\item{warning}{[Output Only] The warning that replaces the list of managed instance groups when the list is empty}
}
\value{
InstanceGroupManagersScopedList object
}
\description{
InstanceGroupManagersScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersScopedList functions: \code{\link{InstanceGroupManagersScopedList.warning.data}},
  \code{\link{InstanceGroupManagersScopedList.warning}}
}
