% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersSetInstanceTemplateRequest}
\alias{InstanceGroupManagersSetInstanceTemplateRequest}
\title{InstanceGroupManagersSetInstanceTemplateRequest Object}
\usage{
InstanceGroupManagersSetInstanceTemplateRequest(instanceTemplate = NULL)
}
\arguments{
\item{instanceTemplate}{The URL of the instance template that is specified for this managed instance group}
}
\value{
InstanceGroupManagersSetInstanceTemplateRequest object
}
\description{
InstanceGroupManagersSetInstanceTemplateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersSetInstanceTemplateRequest functions: \code{\link{instanceGroupManagers.setInstanceTemplate}}
}
