% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagersSetTargetPoolsRequest}
\alias{InstanceGroupManagersSetTargetPoolsRequest}
\title{InstanceGroupManagersSetTargetPoolsRequest Object}
\usage{
InstanceGroupManagersSetTargetPoolsRequest(fingerprint = NULL,
  targetPools = NULL)
}
\arguments{
\item{fingerprint}{The fingerprint of the target pools information}

\item{targetPools}{The list of target pool URLs that instances in this managed instance group belong to}
}
\value{
InstanceGroupManagersSetTargetPoolsRequest object
}
\description{
InstanceGroupManagersSetTargetPoolsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupManagersSetTargetPoolsRequest functions: \code{\link{instanceGroupManagers.setTargetPools}}
}

