% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerUpdatePolicy}
\alias{InstanceGroupManagerUpdatePolicy}
\title{InstanceGroupManagerUpdatePolicy Object}
\usage{
InstanceGroupManagerUpdatePolicy(maxSurge = NULL, maxUnavailable = NULL,
  minReadySec = NULL, minimalAction = NULL, type = NULL)
}
\arguments{
\item{maxSurge}{Maximum number of instances that can be created above the InstanceGroupManager}

\item{maxUnavailable}{Maximum number of instances that can be unavailable during the update process}

\item{minReadySec}{Minimum number of seconds to wait for after a newly created instance becomes available}

\item{minimalAction}{Minimal action to be taken on an instance}

\item{type}{No description}
}
\value{
InstanceGroupManagerUpdatePolicy object
}
\description{
InstanceGroupManagerUpdatePolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
