% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupManagerVersion}
\alias{InstanceGroupManagerVersion}
\title{InstanceGroupManagerVersion Object}
\usage{
InstanceGroupManagerVersion(instanceTemplate = NULL, name = NULL,
  tag = NULL, targetSize = NULL)
}
\arguments{
\item{instanceTemplate}{No description}

\item{name}{Name of the version}

\item{tag}{Tag describing the version}

\item{targetSize}{Intended number of instances that are created from instanceTemplate}
}
\value{
InstanceGroupManagerVersion object
}
\description{
InstanceGroupManagerVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
