% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instanceGroups.removeInstances}
\alias{instanceGroups.removeInstances}
\title{Removes one or more instances from the specified instance group, but does not delete those instances.}
\usage{
instanceGroups.removeInstances(InstanceGroupsRemoveInstancesRequest, project,
  zone, instanceGroup)
}
\arguments{
\item{InstanceGroupsRemoveInstancesRequest}{The \link{InstanceGroupsRemoveInstancesRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone where the instance group is located}

\item{instanceGroup}{The name of the instance group where the specified instances will be removed}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceGroupsRemoveInstancesRequest functions: \code{\link{InstanceGroupsRemoveInstancesRequest}}
}

