% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupsAddInstancesRequest}
\alias{InstanceGroupsAddInstancesRequest}
\title{InstanceGroupsAddInstancesRequest Object}
\usage{
InstanceGroupsAddInstancesRequest(instances = NULL)
}
\arguments{
\item{instances}{The list of instances to add to the instance group}
}
\value{
InstanceGroupsAddInstancesRequest object
}
\description{
InstanceGroupsAddInstancesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstanceGroupsAddInstancesRequest functions: \code{\link{instanceGroups.addInstances}}
}
