% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceGroupsScopedList.warning}
\alias{InstanceGroupsScopedList.warning}
\title{InstanceGroupsScopedList.warning Object}
\usage{

  InstanceGroupsScopedList.warning(InstanceGroupsScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{InstanceGroupsScopedList.warning.data}{The \link{InstanceGroupsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
InstanceGroupsScopedList.warning object
}
\description{
InstanceGroupsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] An informational warning that replaces the list of instance groups when the list is empty.
}
\seealso{
Other InstanceGroupsScopedList functions: \code{\link{InstanceGroupsScopedList.warning.data}},
  \code{\link{InstanceGroupsScopedList}}
}

