% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceProperties.labels}
\alias{InstanceProperties.labels}
\title{InstanceProperties.labels Object}
\usage{
InstanceProperties.labels()
}
\value{
InstanceProperties.labels object
}
\description{
InstanceProperties.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to instances that are created from this template.
}
\seealso{
Other InstanceProperties functions: \code{\link{InstanceProperties}}
}
