% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instances.attachDisk}
\alias{instances.attachDisk}
\title{Attaches a Disk resource to an instance.}
\usage{
instances.attachDisk(AttachedDisk, project, zone, instance,
  forceAttach = NULL, requestId = NULL)
}
\arguments{
\item{AttachedDisk}{The \link{AttachedDisk} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{The instance name for this request}

\item{forceAttach}{Whether to force attach the disk even if it's currently attached to another instance}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other AttachedDisk functions: \code{\link{AttachedDisk}}
}
