% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instances.startWithEncryptionKey}
\alias{instances.startWithEncryptionKey}
\title{Starts an instance that was stopped using the using the instances().stop method. For more information, see Restart an instance.}
\usage{
instances.startWithEncryptionKey(InstancesStartWithEncryptionKeyRequest,
  project, zone, instance, requestId = NULL)
}
\arguments{
\item{InstancesStartWithEncryptionKeyRequest}{The \link{InstancesStartWithEncryptionKeyRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{Name of the instance resource to start}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstancesStartWithEncryptionKeyRequest functions: \code{\link{InstancesStartWithEncryptionKeyRequest}}
}
