% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{instances.updateAccessConfig}
\alias{instances.updateAccessConfig}
\title{Updates the specified access config from an instance's network interface with the data included in the request.}
\usage{
instances.updateAccessConfig(AccessConfig, project, zone, instance,
  networkInterface, requestId = NULL)
}
\arguments{
\item{AccessConfig}{The \link{AccessConfig} object to pass to this method}

\item{project}{Project ID for this request}

\item{zone}{The name of the zone for this request}

\item{instance}{The instance name for this request}

\item{networkInterface}{The name of the network interface where the access config is attached}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other AccessConfig functions: \code{\link{AccessConfig}},
  \code{\link{instances.addAccessConfig}}
}
