% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetLabelsRequest.labels}
\alias{InstancesSetLabelsRequest.labels}
\title{InstancesSetLabelsRequest.labels Object}
\usage{
InstancesSetLabelsRequest.labels()
}
\value{
InstancesSetLabelsRequest.labels object
}
\description{
InstancesSetLabelsRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of labels to apply for this instance. Changing instance labels will also change the instance tags.Each label key & value must comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash. For example, 'webserver-frontend': 'images'. A label value can also be empty (e.g. 'my-label': '').
}
\seealso{
Other InstancesSetLabelsRequest functions: \code{\link{InstancesSetLabelsRequest}},
  \code{\link{instances.setLabels}}
}

