% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetLabelsRequest}
\alias{InstancesSetLabelsRequest}
\title{InstancesSetLabelsRequest Object}
\usage{
InstancesSetLabelsRequest(InstancesSetLabelsRequest.labels = NULL,
  labelFingerprint = NULL, labels = NULL)
}
\arguments{
\item{InstancesSetLabelsRequest.labels}{The \link{InstancesSetLabelsRequest.labels} object or list of objects}

\item{labelFingerprint}{Fingerprint of the previous set of labels for this resource, used to prevent conflicts}

\item{labels}{A list of labels to apply for this instance}
}
\value{
InstancesSetLabelsRequest object
}
\description{
InstancesSetLabelsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetLabelsRequest functions: \code{\link{InstancesSetLabelsRequest.labels}},
  \code{\link{instances.setLabels}}
}

