% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetMachineResourcesRequest}
\alias{InstancesSetMachineResourcesRequest}
\title{InstancesSetMachineResourcesRequest Object}
\usage{
InstancesSetMachineResourcesRequest(guestAccelerators = NULL)
}
\arguments{
\item{guestAccelerators}{No description}
}
\value{
InstancesSetMachineResourcesRequest object
}
\description{
InstancesSetMachineResourcesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetMachineResourcesRequest functions: \code{\link{instances.setMachineResources}}
}
