% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetMinCpuPlatformRequest}
\alias{InstancesSetMinCpuPlatformRequest}
\title{InstancesSetMinCpuPlatformRequest Object}
\usage{
InstancesSetMinCpuPlatformRequest(minCpuPlatform = NULL)
}
\arguments{
\item{minCpuPlatform}{Minimum cpu/platform this instance should be started at}
}
\value{
InstancesSetMinCpuPlatformRequest object
}
\description{
InstancesSetMinCpuPlatformRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetMinCpuPlatformRequest functions: \code{\link{instances.setMinCpuPlatform}}
}
