% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesSetServiceAccountRequest}
\alias{InstancesSetServiceAccountRequest}
\title{InstancesSetServiceAccountRequest Object}
\usage{
InstancesSetServiceAccountRequest(email = NULL, scopes = NULL)
}
\arguments{
\item{email}{Email address of the service account}

\item{scopes}{The list of scopes to be made available for this service account}
}
\value{
InstancesSetServiceAccountRequest object
}
\description{
InstancesSetServiceAccountRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesSetServiceAccountRequest functions: \code{\link{instances.setServiceAccount}}
}

