% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstancesStartWithEncryptionKeyRequest}
\alias{InstancesStartWithEncryptionKeyRequest}
\title{InstancesStartWithEncryptionKeyRequest Object}
\usage{
InstancesStartWithEncryptionKeyRequest(disks = NULL,
  instanceEncryptionKey = NULL)
}
\arguments{
\item{disks}{Array of disks associated with this instance that are protected with a customer-supplied encryption key}

\item{instanceEncryptionKey}{Decrypts data associated with an instance that is protected with a customer-supplied encryption key}
}
\value{
InstancesStartWithEncryptionKeyRequest object
}
\description{
InstancesStartWithEncryptionKeyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InstancesStartWithEncryptionKeyRequest functions: \code{\link{instances.startWithEncryptionKey}}
}
