% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceTemplate}
\alias{InstanceTemplate}
\title{InstanceTemplate Object}
\usage{
InstanceTemplate(creationTimestamp = NULL, description = NULL, id = NULL,
  name = NULL, properties = NULL, selfLink = NULL,
  sourceInstance = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] The creation timestamp for this instance template in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] A unique identifier for this instance template}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{properties}{The instance properties for this instance template}

\item{selfLink}{[Output Only] The URL for this instance template}

\item{sourceInstance}{The source instance used to create the template}
}
\value{
InstanceTemplate object
}
\description{
InstanceTemplate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Instance Template resource.
}
\seealso{
Other InstanceTemplate functions: \code{\link{instanceTemplates.insert}}
}
