% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{InstanceTemplateList}
\alias{InstanceTemplateList}
\title{InstanceTemplateList Object}
\usage{
InstanceTemplateList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] A unique identifier for this instance template}

\item{items}{[Output Only] list of InstanceTemplate resources}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] The URL for this instance template list}
}
\value{
InstanceTemplateList object
}
\description{
InstanceTemplateList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of instance templates.
}

