% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{License}
\alias{License}
\title{License Object}
\usage{
License(chargesUseFee = NULL, licenseCode = NULL, name = NULL,
  resourceRequirements = NULL, selfLink = NULL, transferable = NULL)
}
\arguments{
\item{chargesUseFee}{[Output Only] If true, the customer will be charged license fee for running software that contains this license on an instance}

\item{licenseCode}{[Output Only] The unique code used to attach this license to images, snapshots, and disks}

\item{name}{[Output Only] Name of the resource}

\item{resourceRequirements}{No description}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{transferable}{If false, licenses will not be copied from the source resource when creating an image from a disk, disk from snapshot, or snapshot from disk}
}
\value{
License object
}
\description{
License Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A license resource.
}
