% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{LicenseResourceRequirements}
\alias{LicenseResourceRequirements}
\title{LicenseResourceRequirements Object}
\usage{
LicenseResourceRequirements(minGuestCpuCount = NULL, minMemoryMb = NULL)
}
\arguments{
\item{minGuestCpuCount}{Minimum number of guest cpus required to use the Instance}

\item{minMemoryMb}{Minimum memory required to use the Instance}
}
\value{
LicenseResourceRequirements object
}
\description{
LicenseResourceRequirements Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
