% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{MachineTypesScopedList}
\alias{MachineTypesScopedList}
\title{MachineTypesScopedList Object}
\usage{
MachineTypesScopedList(MachineTypesScopedList.warning = NULL,
  MachineTypesScopedList.warning.data = NULL, machineTypes = NULL,
  warning = NULL)
}
\arguments{
\item{MachineTypesScopedList.warning}{The \link{MachineTypesScopedList.warning} object or list of objects}

\item{MachineTypesScopedList.warning.data}{The \link{MachineTypesScopedList.warning.data} object or list of objects}

\item{machineTypes}{[Output Only] List of machine types contained in this scope}

\item{warning}{[Output Only] An informational warning that appears when the machine types list is empty}
}
\value{
MachineTypesScopedList object
}
\description{
MachineTypesScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MachineTypesScopedList functions: \code{\link{MachineTypesScopedList.warning.data}},
  \code{\link{MachineTypesScopedList.warning}}
}
