% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ManagedInstance}
\alias{ManagedInstance}
\title{ManagedInstance Object}
\usage{
ManagedInstance(currentAction = NULL, id = NULL, instance = NULL,
  instanceStatus = NULL, instanceTemplate = NULL, lastAttempt = NULL,
  standbyMode = NULL, tag = NULL)
}
\arguments{
\item{currentAction}{[Output Only] The current action that the managed instance group has scheduled for the instance}

\item{id}{[Output only] The unique identifier for this resource}

\item{instance}{[Output Only] The URL of the instance}

\item{instanceStatus}{[Output Only] The status of the instance}

\item{instanceTemplate}{[Output Only] The intended template of the instance}

\item{lastAttempt}{[Output Only] Information about the last attempt to create or delete the instance}

\item{standbyMode}{[Output Only] Standby mode of the instance}

\item{tag}{[Output Only] Tag describing the version}
}
\value{
ManagedInstance object
}
\description{
ManagedInstance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

