% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ManagedInstanceLastAttempt}
\alias{ManagedInstanceLastAttempt}
\title{ManagedInstanceLastAttempt Object}
\usage{
ManagedInstanceLastAttempt(ManagedInstanceLastAttempt.errors = NULL,
  ManagedInstanceLastAttempt.errors.errors = NULL, errors = NULL)
}
\arguments{
\item{ManagedInstanceLastAttempt.errors}{The \link{ManagedInstanceLastAttempt.errors} object or list of objects}

\item{ManagedInstanceLastAttempt.errors.errors}{The \link{ManagedInstanceLastAttempt.errors.errors} object or list of objects}

\item{errors}{[Output Only] Encountered errors during the last attempt to create or delete the instance}
}
\value{
ManagedInstanceLastAttempt object
}
\description{
ManagedInstanceLastAttempt Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ManagedInstanceLastAttempt functions: \code{\link{ManagedInstanceLastAttempt.errors.errors}},
  \code{\link{ManagedInstanceLastAttempt.errors}}
}
