% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ManagedInstanceVersion}
\alias{ManagedInstanceVersion}
\title{ManagedInstanceVersion Object}
\usage{
ManagedInstanceVersion(instanceTemplate = NULL, name = NULL)
}
\arguments{
\item{instanceTemplate}{[Output Only] The intended template of the instance}

\item{name}{[Output Only] Name of the version}
}
\value{
ManagedInstanceVersion object
}
\description{
ManagedInstanceVersion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
