% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Metadata}
\alias{Metadata}
\title{Metadata Object}
\usage{
Metadata(Metadata.items = NULL, fingerprint = NULL, items = NULL)
}
\arguments{
\item{Metadata.items}{The \link{Metadata.items} object or list of objects}

\item{fingerprint}{Specifies a fingerprint for this request, which is essentially a hash of the metadata's contents and used for optimistic locking}

\item{items}{Array of key/value pairs}
}
\value{
Metadata object
}
\description{
Metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A metadata key/value entry.
}
\seealso{
Other Metadata functions: \code{\link{Metadata.items}},
  \code{\link{instances.setMetadata}},
  \code{\link{projects.setCommonInstanceMetadata}}
}
