% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Network}
\alias{Network}
\title{Network Object}
\usage{
Network(IPv4Range = NULL, autoCreateSubnetworks = NULL,
  creationTimestamp = NULL, crossVmEncryption = NULL, description = NULL,
  gatewayIPv4 = NULL, id = NULL, loadBalancerVmEncryption = NULL,
  name = NULL, peerings = NULL, selfLink = NULL, subnetworks = NULL)
}
\arguments{
\item{IPv4Range}{The range of internal addresses that are legal on this network}

\item{autoCreateSubnetworks}{When set to true, the network is created in 'auto subnet mode'}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{crossVmEncryption}{[Output Only] Type of VM-to-VM traffic encryption for this network}

\item{description}{An optional description of this resource}

\item{gatewayIPv4}{A gateway address for default routing to other networks}

\item{id}{[Output Only] The unique identifier for the resource}

\item{loadBalancerVmEncryption}{[Output Only] Type of LB-to-VM traffic encryption for this network}

\item{name}{Name of the resource}

\item{peerings}{[Output Only] List of network peerings for the resource}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{subnetworks}{[Output Only] Server-defined fully-qualified URLs for all subnetworks in this network}
}
\value{
Network object
}
\description{
Network Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Network resource. Read Networks and Firewalls for more information.
}
\seealso{
Other Network functions: \code{\link{networks.insert}}
}
