% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{networks.addPeering}
\alias{networks.addPeering}
\title{Adds a peering to the specified network.}
\usage{
networks.addPeering(NetworksAddPeeringRequest, project, network,
  requestId = NULL)
}
\arguments{
\item{NetworksAddPeeringRequest}{The \link{NetworksAddPeeringRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{network}{Name of the network resource to add peering to}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other NetworksAddPeeringRequest functions: \code{\link{NetworksAddPeeringRequest}}
}
