% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{NetworksAddPeeringRequest}
\alias{NetworksAddPeeringRequest}
\title{NetworksAddPeeringRequest Object}
\usage{
NetworksAddPeeringRequest(autoCreateRoutes = NULL, name = NULL,
  peerNetwork = NULL)
}
\arguments{
\item{autoCreateRoutes}{Whether Google Compute Engine manages the routes automatically}

\item{name}{Name of the peering, which should conform to RFC1035}

\item{peerNetwork}{URL of the peer network}
}
\value{
NetworksAddPeeringRequest object
}
\description{
NetworksAddPeeringRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other NetworksAddPeeringRequest functions: \code{\link{networks.addPeering}}
}
