% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{OperationsScopedList.warning}
\alias{OperationsScopedList.warning}
\title{OperationsScopedList.warning Object}
\usage{
OperationsScopedList.warning(OperationsScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{OperationsScopedList.warning.data}{The \link{OperationsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
OperationsScopedList.warning object
}
\description{
OperationsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of operations when the list is empty.
}
\seealso{
Other OperationsScopedList functions: \code{\link{OperationsScopedList.warning.data}},
  \code{\link{OperationsScopedList}}
}
