% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{projects.getXpnResources}
\alias{projects.getXpnResources}
\title{Get XPN resources associated with this host project.}
\usage{
projects.getXpnResources(project, filter = NULL, maxResults = NULL,
  order_by = NULL, pageToken = NULL)
}
\arguments{
\item{project}{Project ID for this request}

\item{filter}{}

\item{maxResults}{}

\item{order_by}{}

\item{pageToken}{}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
