% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{projects.moveInstance}
\alias{projects.moveInstance}
\title{Moves an instance and its attached persistent disks from one zone to another.}
\usage{
projects.moveInstance(InstanceMoveRequest, project, requestId = NULL)
}
\arguments{
\item{InstanceMoveRequest}{The \link{InstanceMoveRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other InstanceMoveRequest functions: \code{\link{InstanceMoveRequest}}
}
