% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ProjectsDisableXpnResourceRequest}
\alias{ProjectsDisableXpnResourceRequest}
\title{ProjectsDisableXpnResourceRequest Object}
\usage{
ProjectsDisableXpnResourceRequest(xpnResource = NULL)
}
\arguments{
\item{xpnResource}{XPN resource ID}
}
\value{
ProjectsDisableXpnResourceRequest object
}
\description{
ProjectsDisableXpnResourceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProjectsDisableXpnResourceRequest functions: \code{\link{projects.disableXpnResource}}
}

