% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ProjectsEnableXpnResourceRequest}
\alias{ProjectsEnableXpnResourceRequest}
\title{ProjectsEnableXpnResourceRequest Object}
\usage{
ProjectsEnableXpnResourceRequest(xpnResource = NULL)
}
\arguments{
\item{xpnResource}{XPN resource ID}
}
\value{
ProjectsEnableXpnResourceRequest object
}
\description{
ProjectsEnableXpnResourceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProjectsEnableXpnResourceRequest functions: \code{\link{projects.enableXpnResource}}
}
