% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ProjectsGetXpnResources}
\alias{ProjectsGetXpnResources}
\title{ProjectsGetXpnResources Object}
\usage{
ProjectsGetXpnResources(resources = NULL)
}
\arguments{
\item{resources}{XPN resources attached to this project as their XPN host}
}
\value{
ProjectsGetXpnResources object
}
\description{
ProjectsGetXpnResources Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

