% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ProjectsSetDefaultServiceAccountRequest}
\alias{ProjectsSetDefaultServiceAccountRequest}
\title{ProjectsSetDefaultServiceAccountRequest Object}
\usage{
ProjectsSetDefaultServiceAccountRequest(email = NULL)
}
\arguments{
\item{email}{Email address of the service account}
}
\value{
ProjectsSetDefaultServiceAccountRequest object
}
\description{
ProjectsSetDefaultServiceAccountRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProjectsSetDefaultServiceAccountRequest functions: \code{\link{projects.setDefaultServiceAccount}}
}
