% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Region}
\alias{Region}
\title{Region Object}
\usage{
Region(creationTimestamp = NULL, deprecated = NULL, description = NULL,
  id = NULL, name = NULL, quotas = NULL, selfLink = NULL,
  status = NULL, zones = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{[Output Only] The deprecation status associated with this region}

\item{description}{[Output Only] Textual description of the resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{[Output Only] Name of the resource}

\item{quotas}{[Output Only] Quotas assigned to this region}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] Status of the region, either UP or DOWN}

\item{zones}{[Output Only] A list of zones available in this region, in the form of resource URLs}
}
\value{
Region object
}
\description{
Region Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Region resource.
}

