% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionBackendServices.insert}
\alias{regionBackendServices.insert}
\title{Creates a regional BackendService resource in the specified project using the data included in the request. There are several restrictions and guidelines to keep in mind when creating a regional backend service. Read  Restrictions and Guidelines for more information.}
\usage{
regionBackendServices.insert(BackendService, project, region)
}
\arguments{
\item{BackendService}{The \link{BackendService} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other BackendService functions: \code{\link{BackendService}},
  \code{\link{backendServices.insert}},
  \code{\link{backendServices.patch}},
  \code{\link{backendServices.update}},
  \code{\link{regionBackendServices.patch}},
  \code{\link{regionBackendServices.update}}
}

