% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionDisks.resize}
\alias{regionDisks.resize}
\title{Resizes the specified regional persistent disk.}
\usage{
regionDisks.resize(RegionDisksResizeRequest, project, region, disk)
}
\arguments{
\item{RegionDisksResizeRequest}{The \link{RegionDisksResizeRequest} object to pass to this method}

\item{project}{The project ID for this request}

\item{region}{Name of the region for this request}

\item{disk}{Name of the regional persistent disk}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other RegionDisksResizeRequest functions: \code{\link{RegionDisksResizeRequest}}
}

