% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionDisksResizeRequest}
\alias{RegionDisksResizeRequest}
\title{RegionDisksResizeRequest Object}
\usage{
RegionDisksResizeRequest(sizeGb = NULL)
}
\arguments{
\item{sizeGb}{The new size of the regional persistent disk, which is specified in GB}
}
\value{
RegionDisksResizeRequest object
}
\description{
RegionDisksResizeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionDisksResizeRequest functions: \code{\link{regionDisks.resize}}
}
