% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroupManagers.resize}
\alias{regionInstanceGroupManagers.resize}
\title{Changes the intended size for the managed instance group. If you increase the size, the group schedules actions to create new instances using the current instance template. If you decrease the size, the group schedules delete actions on one or more instances. The resize operation is marked DONE when the resize actions are scheduled even if the group has not yet added or deleted any instances. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method.}
\usage{
regionInstanceGroupManagers.resize(project, region, instanceGroupManager, size)
}
\arguments{
\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{instanceGroupManager}{Name of the managed instance group}

\item{size}{Number of instances that should exist in this instance group manager}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}

