% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{regionInstanceGroupManagers.setTargetPools}
\alias{regionInstanceGroupManagers.setTargetPools}
\title{Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.}
\usage{

  regionInstanceGroupManagers.setTargetPools(RegionInstanceGroupManagersSetTargetPoolsRequest,
  project, region, instanceGroupManager, requestId = NULL)
}
\arguments{
\item{RegionInstanceGroupManagersSetTargetPoolsRequest}{The \link{RegionInstanceGroupManagersSetTargetPoolsRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{instanceGroupManager}{Name of the managed instance group}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other RegionInstanceGroupManagersSetTargetPoolsRequest functions: \code{\link{RegionInstanceGroupManagersSetTargetPoolsRequest}}
}
