% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionSetLabelsRequest.labels}
\alias{RegionSetLabelsRequest.labels}
\title{RegionSetLabelsRequest.labels Object}
\usage{
RegionSetLabelsRequest.labels()
}
\value{
RegionSetLabelsRequest.labels object
}
\description{
RegionSetLabelsRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels to set for this resource.
}
\seealso{
Other RegionSetLabelsRequest functions: \code{\link{RegionSetLabelsRequest}},
  \code{\link{regionDisks.setLabels}}
}

