% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RegionSetLabelsRequest}
\alias{RegionSetLabelsRequest}
\title{RegionSetLabelsRequest Object}
\usage{
RegionSetLabelsRequest(RegionSetLabelsRequest.labels = NULL,
  labelFingerprint = NULL, labels = NULL)
}
\arguments{
\item{RegionSetLabelsRequest.labels}{The \link{RegionSetLabelsRequest.labels} object or list of objects}

\item{labelFingerprint}{The fingerprint of the previous set of labels for this resource, used to detect conflicts}

\item{labels}{The labels to set for this resource}
}
\value{
RegionSetLabelsRequest object
}
\description{
RegionSetLabelsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other RegionSetLabelsRequest functions: \code{\link{RegionSetLabelsRequest.labels}},
  \code{\link{regionDisks.setLabels}}
}

