% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ResourceCommitment}
\alias{ResourceCommitment}
\title{ResourceCommitment Object}
\usage{
ResourceCommitment(amount = NULL, type = NULL)
}
\arguments{
\item{amount}{The amount of the resource purchased (in a type-dependent unit, such as bytes)}

\item{type}{Type of resource for which this commitment applies}
}
\value{
ResourceCommitment object
}
\description{
ResourceCommitment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Commitment for a particular resource (a Commitment is composed of one or more of these).
}
