% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ResourceGroupReference}
\alias{ResourceGroupReference}
\title{ResourceGroupReference Object}
\usage{
ResourceGroupReference(group = NULL)
}
\arguments{
\item{group}{A URI referencing one of the resource views listed in the backend service}
}
\value{
ResourceGroupReference object
}
\description{
ResourceGroupReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ResourceGroupReference functions: \code{\link{backendServices.getHealth}},
  \code{\link{regionBackendServices.getHealth}}
}

