% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Route}
\alias{Route}
\title{Route Object}
\usage{
Route(Route.warnings = NULL, Route.warnings.data = NULL,
  creationTimestamp = NULL, description = NULL, destRange = NULL,
  id = NULL, name = NULL, network = NULL, nextHopGateway = NULL,
  nextHopInstance = NULL, nextHopIp = NULL, nextHopNetwork = NULL,
  nextHopVpnTunnel = NULL, priority = NULL, selfLink = NULL,
  tags = NULL, warnings = NULL)
}
\arguments{
\item{Route.warnings}{The \link{Route.warnings} object or list of objects}

\item{Route.warnings.data}{The \link{Route.warnings.data} object or list of objects}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{destRange}{The destination range of outgoing packets that this route applies to}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{network}{Fully-qualified URL of the network that this route applies to}

\item{nextHopGateway}{The URL to a gateway that should handle matching packets}

\item{nextHopInstance}{The URL to an instance that should handle matching packets}

\item{nextHopIp}{The network IP address of an instance that should handle matching packets}

\item{nextHopNetwork}{The URL of the local network if it should handle matching packets}

\item{nextHopVpnTunnel}{The URL to a VpnTunnel that should handle matching packets}

\item{priority}{The priority of this route}

\item{selfLink}{[Output Only] Server-defined fully-qualified URL for this resource}

\item{tags}{A list of instance tags to which this route applies}

\item{warnings}{[Output Only] If potential misconfigurations are detected for this route, this field will be populated with warning messages}
}
\value{
Route object
}
\description{
Route Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Route resource. A route specifies how certain packets should be handled by the network. Routes are associated with instances by tags and the set of routes for a particular instance is called its routing table.For each packet leaving a instance, the system searches that instance's routing table for a single best matching route. Routes match packets by destination IP address, preferring smaller or more specific ranges over larger ones. If there is a tie, the system selects the route with the smallest priority value. If there is still a tie, it uses the layer three and four packet headers to select just one of the remaining matching routes. The packet is then forwarded as specified by the nextHop field of the winning route - either to another instance destination, a instance gateway or a Google Compute Engine-operated gateway.Packets that do not match any route in the sending instance's routing table are dropped.
}
\seealso{
Other Route functions: \code{\link{Route.warnings.data}},
  \code{\link{Route.warnings}}, \code{\link{routes.insert}}
}

