% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Route.warnings}
\alias{Route.warnings}
\title{Route.warnings Object}
\usage{
Route.warnings(Route.warnings.data = NULL)
}
\arguments{
\item{Route.warnings.data}{The \link{Route.warnings.data} object or list of objects}
}
\value{
Route.warnings object
}
\description{
Route.warnings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Route functions: \code{\link{Route.warnings.data}},
  \code{\link{Route}}, \code{\link{routes.insert}}
}

