% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouteList}
\alias{RouteList}
\title{RouteList Object}
\usage{
RouteList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource}

\item{items}{[Output Only] A list of Route resources}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
RouteList object
}
\description{
RouteList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of Route resources.
}

