% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RouterBgp}
\alias{RouterBgp}
\title{RouterBgp Object}
\usage{
RouterBgp(asn = NULL)
}
\arguments{
\item{asn}{Local BGP Autonomous System Number (ASN)}
}
\value{
RouterBgp object
}
\description{
RouterBgp Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

