% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{routers.getRouterStatus}
\alias{routers.getRouterStatus}
\title{Retrieves runtime information of the specified router.}
\usage{
routers.getRouterStatus(project, region, router)
}
\arguments{
\item{project}{Project ID for this request}

\item{region}{Name of the region for this request}

\item{router}{Name of the Router resource to query}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
\item https://www.googleapis.com/auth/compute.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute, https://www.googleapis.com/auth/compute.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
