% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{RoutersScopedList.warning}
\alias{RoutersScopedList.warning}
\title{RoutersScopedList.warning Object}
\usage{
RoutersScopedList.warning(RoutersScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{RoutersScopedList.warning.data}{The \link{RoutersScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
RoutersScopedList.warning object
}
\description{
RoutersScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of routers when the list is empty.
}
\seealso{
Other RoutersScopedList functions: \code{\link{RoutersScopedList.warning.data}},
  \code{\link{RoutersScopedList}}
}

