% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Snapshot.labels}
\alias{Snapshot.labels}
\title{Snapshot.labels Object}
\usage{
Snapshot.labels()
}
\value{
Snapshot.labels object
}
\description{
Snapshot.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this snapshot. These can be later modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.
}
\seealso{
Other Snapshot functions: \code{\link{Snapshot}},
  \code{\link{disks.createSnapshot}},
  \code{\link{regionDisks.createSnapshot}}
}

