% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SpreadingPolicy}
\alias{SpreadingPolicy}
\title{SpreadingPolicy Object}
\usage{
SpreadingPolicy(zones = NULL)
}
\arguments{
\item{zones}{No description}
}
\value{
SpreadingPolicy object
}
\description{
SpreadingPolicy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deprecated in favor of distribution policy.
}
