% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SpreadingPolicyZoneConfiguration}
\alias{SpreadingPolicyZoneConfiguration}
\title{SpreadingPolicyZoneConfiguration Object}
\usage{
SpreadingPolicyZoneConfiguration(zone = NULL)
}
\arguments{
\item{zone}{URL of the zone where managed instance group is spawning instances (for regional resources)}
}
\value{
SpreadingPolicyZoneConfiguration object
}
\description{
SpreadingPolicyZoneConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
