% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SslCertificate}
\alias{SslCertificate}
\title{SslCertificate Object}
\usage{
SslCertificate(certificate = NULL, creationTimestamp = NULL,
  description = NULL, id = NULL, name = NULL, privateKey = NULL,
  selfLink = NULL)
}
\arguments{
\item{certificate}{A local certificate file}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{privateKey}{A write-only private key in PEM format}

\item{selfLink}{[Output only] Server-defined URL for the resource}
}
\value{
SslCertificate object
}
\description{
SslCertificate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An SslCertificate resource. This resource provides a mechanism to upload an SSL key and certificate to the load balancer to serve secure connections from the user.
}
\seealso{
Other SslCertificate functions: \code{\link{sslCertificates.insert}}
}
