% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Subnetwork}
\alias{Subnetwork}
\title{Subnetwork Object}
\usage{
Subnetwork(creationTimestamp = NULL, description = NULL,
  gatewayAddress = NULL, id = NULL, ipCidrRange = NULL, name = NULL,
  network = NULL, privateIpGoogleAccess = NULL, region = NULL,
  secondaryIpRanges = NULL, selfLink = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{gatewayAddress}{[Output Only] The gateway address for default routes to reach destination addresses outside this subnetwork}

\item{id}{[Output Only] The unique identifier for the resource}

\item{ipCidrRange}{The range of internal addresses that are owned by this subnetwork}

\item{name}{The name of the resource, provided by the client when initially creating the resource}

\item{network}{The URL of the network to which this subnetwork belongs, provided by the client when initially creating the subnetwork}

\item{privateIpGoogleAccess}{Whether the VMs in this subnet can access Google services without assigned external IP addresses}

\item{region}{URL of the region where the Subnetwork resides}

\item{secondaryIpRanges}{An array of configurations for secondary IP ranges for VM instances contained in this subnetwork}

\item{selfLink}{[Output Only] Server-defined URL for the resource}
}
\value{
Subnetwork object
}
\description{
Subnetwork Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Subnetwork resource.
}
\seealso{
Other Subnetwork functions: \code{\link{subnetworks.insert}}
}
