% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{subnetworks.setPrivateIpGoogleAccess}
\alias{subnetworks.setPrivateIpGoogleAccess}
\title{Set whether VMs in this subnet can access Google services without assigning external IP addresses through Cloudpath.}
\usage{

  subnetworks.setPrivateIpGoogleAccess(SubnetworksSetPrivateIpGoogleAccessRequest,
  project, region, subnetwork, requestId = NULL)
}
\arguments{
\item{SubnetworksSetPrivateIpGoogleAccessRequest}{The \link{SubnetworksSetPrivateIpGoogleAccessRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}

\item{subnetwork}{Name of the Subnetwork resource}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other SubnetworksSetPrivateIpGoogleAccessRequest functions: \code{\link{SubnetworksSetPrivateIpGoogleAccessRequest}}
}
