% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SubnetworkSecondaryRange}
\alias{SubnetworkSecondaryRange}
\title{SubnetworkSecondaryRange Object}
\usage{
SubnetworkSecondaryRange(ipCidrRange = NULL, rangeName = NULL)
}
\arguments{
\item{ipCidrRange}{The range of IP addresses belonging to this subnetwork secondary range}

\item{rangeName}{The name associated with this subnetwork secondary range, used when adding an alias IP range to a VM instance}
}
\value{
SubnetworkSecondaryRange object
}
\description{
SubnetworkSecondaryRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a secondary IP range of a subnetwork.
}
