% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{SubnetworksScopedList.warning}
\alias{SubnetworksScopedList.warning}
\title{SubnetworksScopedList.warning Object}
\usage{
SubnetworksScopedList.warning(SubnetworksScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{SubnetworksScopedList.warning.data}{The \link{SubnetworksScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
SubnetworksScopedList.warning object
}
\description{
SubnetworksScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An informational warning that appears when the list of addresses is empty.
}
\seealso{
Other SubnetworksScopedList functions: \code{\link{SubnetworksScopedList.warning.data}},
  \code{\link{SubnetworksScopedList}}
}

