% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetHttpProxies.insert}
\alias{targetHttpProxies.insert}
\title{Creates a TargetHttpProxy resource in the specified project using the data included in the request.}
\usage{
targetHttpProxies.insert(TargetHttpProxy, project, requestId = NULL)
}
\arguments{
\item{TargetHttpProxy}{The \link{TargetHttpProxy} object to pass to this method}

\item{project}{Project ID for this request}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetHttpProxy functions: \code{\link{TargetHttpProxy}}
}
