% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetHttpsProxies.setSslCertificates}
\alias{targetHttpsProxies.setSslCertificates}
\title{Replaces SslCertificates for TargetHttpsProxy.}
\usage{

  targetHttpsProxies.setSslCertificates(TargetHttpsProxiesSetSslCertificatesRequest,
  project, targetHttpsProxy)
}
\arguments{
\item{TargetHttpsProxiesSetSslCertificatesRequest}{The \link{TargetHttpsProxiesSetSslCertificatesRequest} object to pass to this method}

\item{project}{Project ID for this request}

\item{targetHttpsProxy}{Name of the TargetHttpsProxy resource to set an SslCertificates resource for}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetHttpsProxiesSetSslCertificatesRequest functions: \code{\link{TargetHttpsProxiesSetSslCertificatesRequest}}
}

