% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolAggregatedList}
\alias{TargetPoolAggregatedList}
\title{TargetPoolAggregatedList Object}
\usage{
TargetPoolAggregatedList(TargetPoolAggregatedList.items = NULL, id = NULL,
  items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{TargetPoolAggregatedList.items}{The \link{TargetPoolAggregatedList.items} object or list of objects}

\item{id}{[Output Only] Unique identifier for the resource}

\item{items}{[Output Only] A map of scoped target pool lists}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
TargetPoolAggregatedList object
}
\description{
TargetPoolAggregatedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolAggregatedList functions: \code{\link{TargetPoolAggregatedList.items}}
}
