% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetPools.insert}
\alias{targetPools.insert}
\title{Creates a target pool in the specified project and region using the data included in the request.}
\usage{
targetPools.insert(TargetPool, project, region)
}
\arguments{
\item{TargetPool}{The \link{TargetPool} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region scoping this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetPool functions: \code{\link{TargetPool}}
}

