% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsAddHealthCheckRequest}
\alias{TargetPoolsAddHealthCheckRequest}
\title{TargetPoolsAddHealthCheckRequest Object}
\usage{
TargetPoolsAddHealthCheckRequest(healthChecks = NULL)
}
\arguments{
\item{healthChecks}{A list of HttpHealthCheck resources to add to the target pool}
}
\value{
TargetPoolsAddHealthCheckRequest object
}
\description{
TargetPoolsAddHealthCheckRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolsAddHealthCheckRequest functions: \code{\link{targetPools.addHealthCheck}}
}

