% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsRemoveHealthCheckRequest}
\alias{TargetPoolsRemoveHealthCheckRequest}
\title{TargetPoolsRemoveHealthCheckRequest Object}
\usage{
TargetPoolsRemoveHealthCheckRequest(healthChecks = NULL)
}
\arguments{
\item{healthChecks}{Health check URL to be removed}
}
\value{
TargetPoolsRemoveHealthCheckRequest object
}
\description{
TargetPoolsRemoveHealthCheckRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetPoolsRemoveHealthCheckRequest functions: \code{\link{targetPools.removeHealthCheck}}
}
