% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetPoolsScopedList.warning}
\alias{TargetPoolsScopedList.warning}
\title{TargetPoolsScopedList.warning Object}
\usage{
TargetPoolsScopedList.warning(TargetPoolsScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{TargetPoolsScopedList.warning.data}{The \link{TargetPoolsScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
TargetPoolsScopedList.warning object
}
\description{
TargetPoolsScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Informational warning which replaces the list of addresses when the list is empty.
}
\seealso{
Other TargetPoolsScopedList functions: \code{\link{TargetPoolsScopedList.warning.data}},
  \code{\link{TargetPoolsScopedList}}
}

