% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetSslProxy}
\alias{TargetSslProxy}
\title{TargetSslProxy Object}
\usage{
TargetSslProxy(clientSslPolicy = NULL, creationTimestamp = NULL,
  description = NULL, id = NULL, name = NULL, proxyHeader = NULL,
  selfLink = NULL, service = NULL, sslCertificates = NULL)
}
\arguments{
\item{clientSslPolicy}{URL to ClientSslPolicy resource which controls the set of allowed SSL versions and ciphers}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{proxyHeader}{Specifies the type of proxy header to append before sending data to the backend, either NONE or PROXY_V1}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{service}{URL to the BackendService resource}

\item{sslCertificates}{URLs to SslCertificate resources that are used to authenticate connections to Backends}
}
\value{
TargetSslProxy object
}
\description{
TargetSslProxy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TargetSslProxy resource. This resource defines an SSL proxy.
}
\seealso{
Other TargetSslProxy functions: \code{\link{targetSslProxies.insert}}
}

