% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetTcpProxiesSetBackendServiceRequest}
\alias{TargetTcpProxiesSetBackendServiceRequest}
\title{TargetTcpProxiesSetBackendServiceRequest Object}
\usage{
TargetTcpProxiesSetBackendServiceRequest(service = NULL)
}
\arguments{
\item{service}{The URL of the new BackendService resource for the targetTcpProxy}
}
\value{
TargetTcpProxiesSetBackendServiceRequest object
}
\description{
TargetTcpProxiesSetBackendServiceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetTcpProxiesSetBackendServiceRequest functions: \code{\link{targetTcpProxies.setBackendService}}
}
