% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetTcpProxiesSetProxyHeaderRequest}
\alias{TargetTcpProxiesSetProxyHeaderRequest}
\title{TargetTcpProxiesSetProxyHeaderRequest Object}
\usage{
TargetTcpProxiesSetProxyHeaderRequest(proxyHeader = NULL)
}
\arguments{
\item{proxyHeader}{The new type of proxy header to append before sending data to the backend}
}
\value{
TargetTcpProxiesSetProxyHeaderRequest object
}
\description{
TargetTcpProxiesSetProxyHeaderRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetTcpProxiesSetProxyHeaderRequest functions: \code{\link{targetTcpProxies.setProxyHeader}}
}
