% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetTcpProxy}
\alias{TargetTcpProxy}
\title{TargetTcpProxy Object}
\usage{
TargetTcpProxy(creationTimestamp = NULL, description = NULL, id = NULL,
  name = NULL, proxyHeader = NULL, selfLink = NULL, service = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{proxyHeader}{Specifies the type of proxy header to append before sending data to the backend, either NONE or PROXY_V1}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{service}{URL to the BackendService resource}
}
\value{
TargetTcpProxy object
}
\description{
TargetTcpProxy Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TargetTcpProxy resource. This resource defines a TCP proxy.
}
\seealso{
Other TargetTcpProxy functions: \code{\link{targetTcpProxies.insert}}
}
