% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGateway}
\alias{TargetVpnGateway}
\title{TargetVpnGateway Object}
\usage{
TargetVpnGateway(creationTimestamp = NULL, description = NULL,
  forwardingRules = NULL, id = NULL, name = NULL, network = NULL,
  region = NULL, selfLink = NULL, status = NULL, tunnels = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{forwardingRules}{[Output Only] A list of URLs to the ForwardingRule resources}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{Name of the resource}

\item{network}{URL of the network to which this VPN gateway is attached}

\item{region}{[Output Only] URL of the region where the target VPN gateway resides}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] The status of the VPN gateway}

\item{tunnels}{[Output Only] A list of URLs to VpnTunnel resources}
}
\value{
TargetVpnGateway object
}
\description{
TargetVpnGateway Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Target VPN gateway resource.
}
\seealso{
Other TargetVpnGateway functions: \code{\link{targetVpnGateways.insert}}
}

