% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{targetVpnGateways.insert}
\alias{targetVpnGateways.insert}
\title{Creates a target VPN gateway in the specified project and region using the data included in the request.}
\usage{
targetVpnGateways.insert(TargetVpnGateway, project, region, requestId = NULL)
}
\arguments{
\item{TargetVpnGateway}{The \link{TargetVpnGateway} object to pass to this method}

\item{project}{Project ID for this request}

\item{region}{Name of the region for this request}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other TargetVpnGateway functions: \code{\link{TargetVpnGateway}}
}
