% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGatewaysScopedList}
\alias{TargetVpnGatewaysScopedList}
\title{TargetVpnGatewaysScopedList Object}
\usage{
TargetVpnGatewaysScopedList(TargetVpnGatewaysScopedList.warning = NULL,
  TargetVpnGatewaysScopedList.warning.data = NULL, targetVpnGateways = NULL,
  warning = NULL)
}
\arguments{
\item{TargetVpnGatewaysScopedList.warning}{The \link{TargetVpnGatewaysScopedList.warning} object or list of objects}

\item{TargetVpnGatewaysScopedList.warning.data}{The \link{TargetVpnGatewaysScopedList.warning.data} object or list of objects}

\item{targetVpnGateways}{[Output Only] List of target vpn gateways contained in this scope}

\item{warning}{[Output Only] Informational warning which replaces the list of addresses when the list is empty}
}
\value{
TargetVpnGatewaysScopedList object
}
\description{
TargetVpnGatewaysScopedList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetVpnGatewaysScopedList functions: \code{\link{TargetVpnGatewaysScopedList.warning.data}},
  \code{\link{TargetVpnGatewaysScopedList.warning}}
}

