% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGatewaysScopedList.warning.data}
\alias{TargetVpnGatewaysScopedList.warning.data}
\title{TargetVpnGatewaysScopedList.warning.data Object}
\usage{
TargetVpnGatewaysScopedList.warning.data()
}
\value{
TargetVpnGatewaysScopedList.warning.data object
}
\description{
TargetVpnGatewaysScopedList.warning.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other TargetVpnGatewaysScopedList functions: \code{\link{TargetVpnGatewaysScopedList.warning}},
  \code{\link{TargetVpnGatewaysScopedList}}
}

