% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{TargetVpnGatewaysScopedList.warning}
\alias{TargetVpnGatewaysScopedList.warning}
\title{TargetVpnGatewaysScopedList.warning Object}
\usage{

  TargetVpnGatewaysScopedList.warning(TargetVpnGatewaysScopedList.warning.data = NULL,
  code = NULL, data = NULL, message = NULL)
}
\arguments{
\item{TargetVpnGatewaysScopedList.warning.data}{The \link{TargetVpnGatewaysScopedList.warning.data} object or list of objects}

\item{code}{[Output Only] A warning code, if applicable}

\item{data}{[Output Only] Metadata about this warning in key: value format}

\item{message}{[Output Only] A human-readable description of the warning code}
}
\value{
TargetVpnGatewaysScopedList.warning object
}
\description{
TargetVpnGatewaysScopedList.warning Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] Informational warning which replaces the list of addresses when the list is empty.
}
\seealso{
Other TargetVpnGatewaysScopedList functions: \code{\link{TargetVpnGatewaysScopedList.warning.data}},
  \code{\link{TargetVpnGatewaysScopedList}}
}

