% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMapList}
\alias{UrlMapList}
\title{UrlMapList Object}
\usage{
UrlMapList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource}

\item{items}{A list of UrlMap resources}

\item{nextPageToken}{[Output Only] This token allows you to get the next page of results for list requests}

\item{selfLink}{[Output Only] Server-defined URL for this resource}
}
\value{
UrlMapList object
}
\description{
UrlMapList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains a list of UrlMap resources.
}
