% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMapReference}
\alias{UrlMapReference}
\title{UrlMapReference Object}
\usage{
UrlMapReference(urlMap = NULL)
}
\arguments{
\item{urlMap}{No description}
}
\value{
UrlMapReference object
}
\description{
UrlMapReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other UrlMapReference functions: \code{\link{targetHttpProxies.setUrlMap}},
  \code{\link{targetHttpsProxies.setUrlMap}}
}

