% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{urlMaps.patch}
\alias{urlMaps.patch}
\title{Updates the specified UrlMap resource with the data included in the request. This method supports patch semantics.}
\usage{
urlMaps.patch(UrlMap, project, urlMap, requestId = NULL)
}
\arguments{
\item{UrlMap}{The \link{UrlMap} object to pass to this method}

\item{project}{Project ID for this request}

\item{urlMap}{Name of the UrlMap resource to update}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}

Other UrlMap functions: \code{\link{UrlMap}},
  \code{\link{urlMaps.insert}},
  \code{\link{urlMaps.update}}
}
