% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UrlMapTest}
\alias{UrlMapTest}
\title{UrlMapTest Object}
\usage{
UrlMapTest(description = NULL, host = NULL, path = NULL, service = NULL)
}
\arguments{
\item{description}{Description of this test case}

\item{host}{Host portion of the URL}

\item{path}{Path portion of the URL}

\item{service}{Expected BackendService resource the given URL should be mapped to}
}
\value{
UrlMapTest object
}
\description{
UrlMapTest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message for the expected URL mappings.
}

