% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{UsageExportLocation}
\alias{UsageExportLocation}
\title{UsageExportLocation Object}
\usage{
UsageExportLocation(bucketName = NULL, reportNamePrefix = NULL)
}
\arguments{
\item{bucketName}{The name of an existing bucket in Cloud Storage where the usage report object is stored}

\item{reportNamePrefix}{An optional prefix for the name of the usage report object stored in bucketName}
}
\value{
UsageExportLocation object
}
\description{
UsageExportLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The location in Cloud Storage and naming method of the daily usage report. Contains bucket_name and report_name prefix.
}
\seealso{
Other UsageExportLocation functions: \code{\link{projects.setUsageExportBucket}}
}

