% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnel.labels}
\alias{VpnTunnel.labels}
\title{VpnTunnel.labels Object}
\usage{
VpnTunnel.labels()
}
\value{
VpnTunnel.labels object
}
\description{
VpnTunnel.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels to apply to this VpnTunnel. These can be later modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.
}
\seealso{
Other VpnTunnel functions: \code{\link{VpnTunnel}},
  \code{\link{vpnTunnels.insert}}
}
