% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnel}
\alias{VpnTunnel}
\title{VpnTunnel Object}
\usage{
VpnTunnel(VpnTunnel.labels = NULL, creationTimestamp = NULL,
  description = NULL, detailedStatus = NULL, id = NULL,
  ikeVersion = NULL, labelFingerprint = NULL, labels = NULL,
  localTrafficSelector = NULL, name = NULL, peerIp = NULL,
  region = NULL, remoteTrafficSelector = NULL, router = NULL,
  selfLink = NULL, sharedSecret = NULL, sharedSecretHash = NULL,
  status = NULL, targetVpnGateway = NULL)
}
\arguments{
\item{VpnTunnel.labels}{The \link{VpnTunnel.labels} object or list of objects}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional description of this resource}

\item{detailedStatus}{[Output Only] Detailed status message for the VPN tunnel}

\item{id}{[Output Only] The unique identifier for the resource}

\item{ikeVersion}{IKE protocol version to use when establishing the VPN tunnel with peer VPN gateway}

\item{labelFingerprint}{A fingerprint for the labels being applied to this VpnTunnel, which is essentially a hash of the labels set used for optimistic locking}

\item{labels}{Labels to apply to this VpnTunnel}

\item{localTrafficSelector}{Local traffic selector to use when establishing the VPN tunnel with peer VPN gateway}

\item{name}{Name of the resource}

\item{peerIp}{IP address of the peer VPN gateway}

\item{region}{[Output Only] URL of the region where the VPN tunnel resides}

\item{remoteTrafficSelector}{Remote traffic selectors to use when establishing the VPN tunnel with peer VPN gateway}

\item{router}{URL of router resource to be used for dynamic routing}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{sharedSecret}{Shared secret used to set the secure session between the Cloud VPN gateway and the peer VPN gateway}

\item{sharedSecretHash}{Hash of the shared secret}

\item{status}{[Output Only] The status of the VPN tunnel}

\item{targetVpnGateway}{URL of the VPN gateway with which this VPN tunnel is associated}
}
\value{
VpnTunnel object
}
\description{
VpnTunnel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other VpnTunnel functions: \code{\link{VpnTunnel.labels}},
  \code{\link{vpnTunnels.insert}}
}
