% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_functions.R
\name{vpnTunnels.delete}
\alias{vpnTunnels.delete}
\title{Deletes the specified VpnTunnel resource.}
\usage{
vpnTunnels.delete(project, region, vpnTunnel, requestId = NULL)
}
\arguments{
\item{project}{Project ID for this request}

\item{region}{Name of the region for this request}

\item{vpnTunnel}{Name of the VpnTunnel resource to delete}

\item{requestId}{begin_interface: MixerMutationRequestBuilder Request ID to support idempotency}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/compute
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/compute)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/compute/docs/reference/latest/}{Google Documentation}
}
