% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{VpnTunnelsScopedList.warning.data}
\alias{VpnTunnelsScopedList.warning.data}
\title{VpnTunnelsScopedList.warning.data Object}
\usage{
VpnTunnelsScopedList.warning.data()
}
\value{
VpnTunnelsScopedList.warning.data object
}
\description{
VpnTunnelsScopedList.warning.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other VpnTunnelsScopedList functions: \code{\link{VpnTunnelsScopedList.warning}},
  \code{\link{VpnTunnelsScopedList}}
}
