% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{XpnResourceId}
\alias{XpnResourceId}
\title{XpnResourceId Object}
\usage{
XpnResourceId(id = NULL, type = NULL)
}
\arguments{
\item{id}{The ID of the XPN resource}

\item{type}{The type of the XPN resource}
}
\value{
XpnResourceId object
}
\description{
XpnResourceId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
XpnResourceId
}

