% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{Zone}
\alias{Zone}
\title{Zone Object}
\usage{
Zone(availableCpuPlatforms = NULL, creationTimestamp = NULL,
  deprecated = NULL, description = NULL, id = NULL, name = NULL,
  region = NULL, selfLink = NULL, status = NULL)
}
\arguments{
\item{availableCpuPlatforms}{[Output Only] Available cpu/platform selections for the zone}

\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{deprecated}{[Output Only] The deprecation status associated with this zone}

\item{description}{[Output Only] Textual description of the resource}

\item{id}{[Output Only] The unique identifier for the resource}

\item{name}{[Output Only] Name of the resource}

\item{region}{[Output Only] Full URL reference to the region which hosts the zone}

\item{selfLink}{[Output Only] Server-defined URL for the resource}

\item{status}{[Output Only] Status of the zone, either UP or DOWN}
}
\value{
Zone object
}
\description{
Zone Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Zone resource.
}
