% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ZoneSetLabelsRequest.labels}
\alias{ZoneSetLabelsRequest.labels}
\title{ZoneSetLabelsRequest.labels Object}
\usage{
ZoneSetLabelsRequest.labels()
}
\value{
ZoneSetLabelsRequest.labels object
}
\description{
ZoneSetLabelsRequest.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The labels to set for this resource.
}
\seealso{
Other ZoneSetLabelsRequest functions: \code{\link{ZoneSetLabelsRequest}},
  \code{\link{disks.setLabels}}
}

