% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{ZoneSetLabelsRequest}
\alias{ZoneSetLabelsRequest}
\title{ZoneSetLabelsRequest Object}
\usage{
ZoneSetLabelsRequest(ZoneSetLabelsRequest.labels = NULL,
  labelFingerprint = NULL, labels = NULL)
}
\arguments{
\item{ZoneSetLabelsRequest.labels}{The \link{ZoneSetLabelsRequest.labels} object or list of objects}

\item{labelFingerprint}{The fingerprint of the previous set of labels for this resource, used to detect conflicts}

\item{labels}{The labels to set for this resource}
}
\value{
ZoneSetLabelsRequest object
}
\description{
ZoneSetLabelsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ZoneSetLabelsRequest functions: \code{\link{ZoneSetLabelsRequest.labels}},
  \code{\link{disks.setLabels}}
}
