% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_objects.R
\name{AccessConfig}
\alias{AccessConfig}
\title{Compute Engine API Objects 
Creates and runs virtual machines on Google Cloud Platform.}
\usage{
AccessConfig(name = NULL, natIP = NULL, type = NULL)
}
\arguments{
\item{name}{Name of this access configuration}

\item{natIP}{An external IP address associated with this instance}

\item{type}{The type of configuration}
}
\value{
AccessConfig object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:01:18
filename: /Users/mark/dev/R/autoGoogleAPI/googlecomputealpha.auto/R/compute_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AccessConfig Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access configuration attached to an instance's network interface. Only one access config per instance is supported.
}
\seealso{
Other AccessConfig functions: \code{\link{instances.addAccessConfig}}
}

